Working with import/export, removed DCP code, and reorganized files into folders
