+++ web_scraper/src/index.ts
@@ -1,34 +1,31 @@
+import { findPath } from './web_scraping/web_scraper';
+import promptSync from 'prompt-sync';
 
-
+/**
+ * Test function
+ * Run node out/web_scraping/index.js
+ * to test web scraping code! :)
+ */
 async function main() {
-    // const findPath = require('./web_scraper.js').findPath;
-    // const startUrl: string = 'https://en.wikipedia.org/wiki/Among_Us';
-    // const endUrl: string = 'https://en.wikipedia.org/wiki/Deductive_reasoning';
-    // const maxChecks: number = 1000;
-    // const maxLayers: number = 5;
-    // const result: string[] = await findPath(startUrl, endUrl, maxChecks, maxLayers);
-    // console.log(result);
     var linksW = { "Gamer": ["Gambling", "Gamer_(disambiguation)", "Video_game", "Video_game#Platform", "Arcade_video_game", "Console_game", "Video_game_console", "Home_video_game_console", "Handheld_game_console", "Electronic_game", "Audio_game", "Handheld_electronic_game", "Online_game", "Browser_game", "Social-network_game", "Mobile_gaming", "PC_game", "Linux_gaming", "Mac_gaming", "Virtual_reality_game", "Video_game_genre", "Action_game", "Beat_%27em_up", "Hack_and_slash", "Fighting_game", "Platform_game", "Shooter_game", "Survival_game", "Battle_royale_game", "Action-adventure_game", "Stealth_game", "Survival_horror", "Adventure_game", "Interactive_fiction", "Interactive_film", "Visual_novel", "Gacha_game", "Role-playing_video_game", "Action_role-playing_game", "Tactical_role-playing_game", "Simulation_video_game", "Construction_and_management_simulation", "Life_simulation_game", "Sports_video_game", "Vehicle_simulation_game", "Strategy_video_game", "4X", "Auto_battler", "Multiplayer_online_battle_arena", "Real-time_strategy", "Real-time_tactics", "Tower_defense", "Turn-based_strategy", "Turn-based_tactics", "Horror_game", "Massively_multiplayer_online_game", "Lists_of_video_games", "List_of_arcade_video_games", "List_of_best-selling_video_games", "List_of_best-selling_video_game_franchises", "List_of_fastest-selling_products", "Arcade_video_game#Highest-grossing", "List_of_highest-grossing_mobile_games", "List_of_longest-running_video_game_franchises", "List_of_most-played_video_games_by_player_count", "List_of_most-played_mobile_games_by_player_count", "List_of_video_games_considered_the_best", "List_of_Game_of_the_Year_awards", "List_of_video_games_notable_for_negative_reception", "Video_game_development", "Artificial_intelligence_in_video_games", "Game_design", "Video_game_design", "Interaction_design", "Video_game_programming", "Video_game_art", "Video_game_graphics", "Video_game_music", "Template_talk:Video_games", "Person", "Player_(game)", "Video_game", "Role-playing_games", "Card_game", "Gambling", "Online_gambling", "Internet_forum", "YouTube", "Twitch_(service)", "Virtual_communities", "Real_life#As_distinct_from_the_Internet", "Social_clubs", "Professional_gamer", "Gambling", "Walsall", "Wargame", "Dungeons_%26_Dragons", "Role-playing_game", "Board_game", "Video_game", "Video_game#Demographics", "Pew_Research_Center", "Women_and_video_games", "Reappropriation", "Learning_curve#Difficulty_curves_in_video_games", "Shigeru_Miyamoto", "Hideo_Kojima", "Mental_toughness", "Olympic_Games", "PlayStation_4", "Stereotype", "Gaymer", "LGBT", "Gay", "Bisexual", "Lesbian", "Transgender", "Video_game", "Warren_Moon", "Marshall_Faulk", "Newbie", "Casual_games", "The_Sims", "Nintendogs", "Target_consumer", "Satoru_Iwata", "Wii_U", "Jon_Radoff", "Disruptor_Beam", "Television", "Ernest_W._Adams", "Scott_Kim", "Professional", "Prize", "Money", "Salaries", "Competitions", "Esports", "South_Korea", "China", "Sponsorship", "United_States_dollar", "Major_League_Gaming", "Halo_2", "Tom_%22Tsquared%22_Taylor", "Team_Final_Boss", "Live_streaming", "Twitch_(service)", "Hitbox", "Mixer.com", "YouTube", "League_of_Legends", "Esports", "Retrogaming", "Retro_games", "Arcade_game", "United_States", "Video_game_console_emulator", "Porting", "Taxonomy_(general)", "Bartle_taxonomy_of_player_types", "Achiever", "Game_badge", "Explorers", "Secret_level", "Glitch", "MDA_framework", "Gameplay", "Jesse_Schell", "Schadenfreude", "Avatar_(computing)", "Player_character", "Pseudonym", "Online_multiplayer", "Electronic_sport", "Middle_name", "Ninja_(video_game_player)", "Sinatraa", "Clan_(video_games)", "Guild", "E-sports", "Performativity", "Graeme_Kirkpatrick", "Gameplay", "The_Escapist_(magazine)", "Cinemaphile", "Gamergate_controversy", "Video_game_journalism", "Demography", "Nintendo_marketing#Demographic_shifts", "Smartphones", "Pc_game", "Hispanics", "African-Americans", "Whites", "Fighting_game", "Quarter_(United_States_coin)", "Gameplay", "Doom_(1993_video_game)", "Call_of_Duty", "Interactive_movie", "The_Sims", "Minecraft", "Counter-Strike", "Leigh_Alexander_(journalist)", "Portal:Video_Games", "Portal:Society", "Entertainment_Consumers_Association", "Gamers_Outreach_Foundation", "Going_Cardboard", "List_of_gaming_topics", "Player_(game)", "Video_game_addiction", "Statista", "ISBN_(identifier)", "ISBN_(identifier)", "Queensland_University_of_Technology", "EurekAlert!", "Frontiers_in_Psychology", "Doi_(identifier)", "PMC_(identifier)", "PMID_(identifier)", "Doi_(identifier)", "ISSN_(identifier)", "S2CID_(identifier)", "Doi_(identifier)", "Hdl_(identifier)", "S2CID_(identifier)", "Washington_Blade", "Wayback_Machine", "Wayback_Machine", "Wayback_Machine", "Institut_national_de_l%27audiovisuel", "Joystiq.com", "Forbes", "Doi_(identifier)", "ISSN_(identifier)", "PMC_(identifier)", "PMID_(identifier)", "Doi_(identifier)", "ISBN_(identifier)", "ISBN_(identifier)", "Doi_(identifier)", "S2CID_(identifier)", "Template_talk:Fandom", "Fandom", "Fantasy_fandom", "Furry_fandom", "Science_fiction_fandom", "Yaoi_fandom", "Anime_and_manga_fandom", "Bibliophilia", "Cinephilia", "Sports_fan", "A_Song_of_Ice_and_Fire_fandom", "Beatlemania", "Deadhead", "Doctor_Who_fandom", "EarthBound_fandom", "Harry_Potter_fandom", "James_Bond_fandom", "Juggalo", "Michael_Jackson_fandom", "Mystery_Science_Theater_3000", "My_Little_Pony:_Friendship_Is_Magic_fandom", "Parrothead", "Shrek_fandom", "Sherlock_Holmes_fandom", "Cultural_impact_of_Star_Wars", "Stargate_fandom", "Steven_Universe#Fandom", "Tolkien_fandom", "Trekkie", "Twilight_fandom", "Anorak_(slang)", "Anti-fan", "Cult_following", "Fan_(person)", "Fursona", "Geek", "Groupie", "Nerd", "Otaku", "Sasaeng_fan", "Shipping_(fandom)", "Stan_(fan)", "Tifosi", "Ultras", "Amateur_press_association", "Car_club", "Boy_racer", "Cruising_(driving)", "Import_scene", "Hot_rod", "Kustom_Kulture", "Lowrider_Club", "Supercar_Season", "Cosplay", "Fan_club", "Fan_convention", "List_of_fandom_nicknames", "Game_club", "Historical_reenactment", "Live_action_role-playing_game", "Motorcycle_club", "Harley_Owners_Group", "Organization_for_Transformative_Works", "Railfan", "Stan_Twitter", "Anime_music_video", "Autograph_collecting", "Doujin", "Doujin_music", "Doujinshi", "Fan_art", "Fan_edit", "Fan_fiction", "Fan_film", "Fan_labor", "Fan_mail", "Fan_translation", "Fanac", "Fangame", "Doujin_soft", "Mod_(video_games)", "Fansite", "Fanspeak", "Fanzine", "Filk_music", "Real_person_fiction", "Vidding", "Zine", "Fan_convention", "Anime_convention", "List_of_anime_conventions", "Comic_book_convention", "List_of_comic_book_conventions", "Doujinshi_convention", "Furry_convention", "Gaming_convention", "List_of_gaming_conventions", "Horror_convention", "List_of_multigenre_conventions", "List_of_professional_wrestling_conventions", "Science_fiction_convention", "List_of_science_fiction_conventions", "List_of_fan_conventions_by_date_of_founding", "Autograph", "Fan_loyalty", "Fan_service", "Odagiri_effect"], "YouTube": ["YouTube_(channel)", "Types_of_business_entity", "Subsidiary", "Online_video_platform", "San_Bruno,_California", "Censorship_of_YouTube", "Organizational_founder", "Chad_Hurley", "Steve_Chen", "Jawed_Karim", "Susan_Wojcicki", "CEO", "Internet", "Video_hosting_service", "Product_(business)", "YouTube_Premium", "YouTube_Music", "YouTube_TV", "YouTube_Kids", "Revenue", "Holding_company", "Google_LLC", "Online_advertising", "Google_AdSense", "Registered_user", "Creative_Commons", "Python_(programming_language)", "C_(programming_language)", "CPython", "C%2B%2B", "Java_(programming_language)", "Guice", "Go_(programming_language)", "JavaScript", "Online_video_platform", "Social_media", "Google", "Steve_Chen", "Chad_Hurley", "Jawed_Karim", "List_of_most_popular_websites", "Google_AdSense", "Mobile_app", "Interoperability", "Music_video", "Video_clip", "News", "Short_film", "Feature_film", "Documentary_film", "Trailer_(promotion)", "Teaser_campaign", "Live_stream", "Vlog", "User-generated_content", "YouTuber", "The_Walt_Disney_Company", "ViacomCBS", "WarnerMedia", "Social_impact_of_YouTube", "Misinformation", "YouTube_copyright_issues", "YouTube_and_privacy", "Censorship_of_YouTube", "Child_protection", "History_of_YouTube", "Chad_Hurley", "Steve_Chen", "Jawed_Karim", "Steve_Chen", "Chad_Hurley", "Jawed_Karim", "PayPal", "EBay", "Indiana_University_of_Pennsylvania", "Computer_science", "University_of_Illinois_at_Urbana%E2%80%93Champaign", "San_Francisco", "Super_Bowl_XXXVIII_halftime_show_controversy", "Janet_Jackson", "Justin_Timberlake", "2004_Indian_Ocean_earthquake_and_tsunami", "Online_dating_service", "Hot_or_Not", "Craigslist", "Venture_capital", "Startup_company", "Sequoia_Capital", "San_Mateo,_California", "Me_at_the_zoo", "San_Diego_Zoo", "Ronaldinho", "Vimeo", "CollegeHumor", "Saturday_Night_Live", "Lazy_Sunday_(The_Lonely_Island_song)", "The_Lonely_Island", "Viral_video", "NBCUniversal", "Universal_Tube_%26_Rollform_Equipment", "Google", "IAC_(company)", "The_Daily_Telegraph", "Bandwidth_(computing)", "Market_share", "ComScore", "San_Bruno,_California", "Salar_Kamangar", "Susan_Wojcicki", "Chief_executive_officer", "Material_Design", "Google_Play_Music", "YouTube_Music", "YouTube_Kids", "Video_gaming", "Amazon.com", "Twitch_(service)", "COVID-19_pandemic", "Stay-at-home_order", "2018_YouTube_headquarters_shooting", "Children%27s_Online_Privacy_Protection_Act", "Preadolescence", "TikTok", "Amazon_Fire_TV", "Roku", "Cyberbullying", "Vote_brigading", "Internet_troll", "YouTube_Rewind_2018", "List_of_most-disliked_YouTube_videos", "VP9", "H.264/MPEG-4_AVC", "Dynamic_Adaptive_Streaming_over_HTTP", "MPEG-4_Part_2", "3GP_and_3G2", "AV1", "Adobe_Flash_Player", "HTML5", "HTML5_video_support", "Dynamic_Adaptive_Streaming_over_HTTP", "Internet_speed", "Data_plan", "Interactive_video", "Widget_(GUI)", "ISNI", "Gigabyte", "Closed_captioning", "Speech_recognition", "Machine_translation", "Digital_container_format", "MPEG-4_Part_14", "Matroska", "Flash_Video", "Audio_Video_Interleave", "WebM", "3GP_and_3G2", "MPEG_program_stream", "QuickTime_File_Format", "Apple_ProRes", "Case-sensitive", "Base64#The_URL_applications", "Uniform_Resource_Locator", "Pixel", "Sorenson_codec", "H.263", "3GP_and_3G2", "720p", "High-definition_video", "4:3", "Aspect_ratio_(image)", "Widescreen", "16:9", "H.264/MPEG-4_AVC", "1080p", "4K_resolution", "4K_resolution", "8K_resolution", "High-dynamic-range_video", "Hybrid_log%E2%80%93gamma", "Perceptual_quantizer", "Rec._2020", "High-frame-rate", "Frames_per_second", "Video_game", "Bitrate", "360-degree_video", "Live_streaming", "U2", "Barack_Obama", "Felix_Baumgartner", "Red_Bull_Stratos", "Public_beta", "Social_media", "GIF", "YouTuber", "Vlogbrothers", "AsapScience", "Lilly_Singh", "MatPat", "Karmin", "The_Key_of_Awesome", "Peter_Hollens", "Rosianna_Halse_Rojas", "Sam_Tsui", "Threadbanger", "Criticism_of_Google#YouTube_user_comments", "Troll_(Internet)", "Time_(magazine)", "Web_2.0", "The_Guardian", "Sexist", "Homophobic", "LOL", "The_Daily_Telegraph", "Comments_section", "The_Huffington_Post", "Offensive_content", "URL", "Misogynistic", "Algorithm", "HTML", "MPEG-4_Part_14", "Plug-in_(computing)", "Google_Checkout", "Cease_and_desist", "Zamzar", "Public_copyright_license", "Creative_Commons", "Smartphone", "Real_Time_Streaming_Protocol", "Apple_Inc.", "H.264/MPEG-4_AVC", "Apple_TV", "IPod_Touch", "IPhone", "HTML5", "GlobalWebIndex", "TiVo", "PlayStation_3", "Wii", "Video_game_console", "Xbox_Live", "Wii_U", "Nintendo_3DS", "Internet_Browser_(Wii_U)", "Roku", "PlayStation_4", "Nintendo_Switch", "Eric_Schmidt", "Internationalization_and_localization", "Hong_Kong", "United_States", "American_English", "Spanish_language_in_the_United_States", "Brazil", "Brazilian_Portuguese", "France", "French_language", "Catalan_language", "Basque_language", "Republic_of_Ireland", "Hiberno-English", "Irish_language", "Italy", "Italian_language", "German_language", "Catalan_language", "Japan", "Japanese_language", "Netherlands", "Dutch_language", "Poland", "Polish_language", "Spain", "Spanish_language", "Galician_language", "Catalan_language", "Basque_language", "United_Kingdom", "British_English", "Mexico", "Mexican_Spanish", "Hong_Kong", "Traditional_Chinese_characters", "Hong_Kong_English", "Taiwan", "Traditional_Chinese_characters", "Australia", "Australian_English", "New_Zealand", "New_Zealand_English", "Canada", "Canadian_English", "Canadian_French", "Germany", "German_language", "Russia", "Russian_language", "South_Korea", "Korean_language", "India", "Hindi_language", "Bengali_language", "Punjabi_language", "Indian_English", "Gujarati_language", "Kannada_language", "Malayalam_language", "Marathi_language", "Tamil_language", "Telugu_language", "Urdu", "Israel", "Hebrew_language", "Arabic_language_in_Israel", "Czech_Republic", "Czech_language", "Sweden", "Swedish_language", "South_Africa", "South_African_English", "Afrikaans_language", "Zulu_language", "Argentina", "Rioplatense_Spanish", "Algeria", "Modern_Standard_Arabic", "French_language", "Egypt", "Modern_Standard_Arabic", "Jordan", "Modern_Standard_Arabic", "Morocco", "French_language", "Modern_Standard_Arabic", "Saudi_Arabia", "Modern_Standard_Arabic", "Tunisia", "Modern_Standard_Arabic", "French_language", "Yemen", "Modern_Standard_Arabic", "Kenya", "English_language", "Swahili_language", "Philippines", "Filipino_language", "Philippine_English", "Singapore", "Singapore_English", "Malay_language", "Simplified_Chinese_characters", "Tamil_language", "Belgium", "Belgian_French", "Dutch_language", "German_language", "Colombia", "Colombian_Spanish", "Uganda", "Ugandan_English", "Swahili_language", "Nigeria", "Nigerian_Standard_English", "Chile", "Spanish_language", "Hungary", "Hungarian_language", "Malaysia", "Malaysian_language", "Malaysian_English", "Peru", "Spanish_language", "United_Arab_Emirates", "Modern_Standard_Arabic", "English_language", "Greece", "Greek_language", "Indonesia", "Indonesian_language", "English_language", "Ghana", "English_language", "Senegal", "French_language", "English_language", "Turkey", "Turkish_language", "Ukraine", "Ukrainian_language", "Denmark", "Danish_language", "Finland", "Finnish_language", "Swedish_language", "Norway", "Norwegian_language", "Switzerland", "German_language", "French_language", "Italian_language", "Austria", "German_language", "Romania", "Romanian_language", "Portugal", "Portuguese_language", "Slovakia", "Slovak_language", "Bahrain", "Modern_Standard_Arabic", "Kuwait", "Modern_Standard_Arabic", "Oman", "Modern_Standard_Arabic", "Qatar", "Modern_Standard_Arabic", "Bosnia_and_Herzegovina", "Bosnian_language", "Croatian_language", "Serbian_language", "Bulgaria", "Bulgarian_language", "Croatia", "Croatian_language", "Estonia", "Estonian_language", "Latvia", "Latvian_language", "Lithuania", "Lithuanian_language", "North_Macedonia", "Macedonian_language", "Serbian_language", "Turkish_language", "Montenegro", "Serbian_language", "Croatian_language", "Serbia", "Serbian_language", "Slovenia", "Slovenian_language", "Thailand", "Thai_language", "Lebanon", "Modern_Standard_Arabic", "Puerto_Rico", "Puerto_Rican_Spanish", "American_English", "Iceland", "Icelandic_language", "Luxembourg", "French_language", "German_language", "Vietnam", "Vietnamese_language", "Libya", "Modern_Standard_Arabic", "Tanzania", "English_language", "Swahili_language", "Zimbabwe", "English_language", "Azerbaijan", "Azerbaijani_language", "Belarus", "Belarusian_language", "Russian_language", "Georgia_(country)", "Georgian_language", "Kazakhstan", "Kazakh_language", "Russian_language", "Iraq", "Modern_Standard_Arabic", "Nepal", "Nepali_language", "Pakistan", "Urdu", "Pakistani_English", "Sri_Lanka", "Sinhala_language", "Tamil_language", "Sri_Lankan_English", "Jamaica", "Jamaican_English", "Malta", "English_language", "Bolivia", "Spanish_language", "Costa_Rica", "Spanish_language", "Ecuador", "Spanish_language", "El_Salvador", "Spanish_language", "Guatemala", "Spanish_language", "Honduras", "Spanish_language", "Nicaragua", "Spanish_language", "Panama", "Spanish_language", "Uruguay", "Spanish_language", "Paraguay", "Spanish_language", "Guarani_language", "Dominican_Republic", "Spanish_language", "Cyprus", "Greek_language", "Turkish_language", "Liechtenstein", "German_language", "Venezuela", "Spanish_language", "Papua_New_Guinea", "English_language", "Bangladesh", "Bengali_language", "English_language", "Turkey", "Mustafa_Kemal_Atat%C3%BCrk", "PRS_for_Music", "Sony_Music_Entertainment", "The_Daily_Dot", "Click_fraud", "Social_Blade", "Vote_brigading", "YouTube_copyright_issues", "Lazy_Sunday_(The_Lonely_Island_song)", "Digital_Millennium_Copyright_Act", "Takedown_notice", "Online_Copyright_Infringement_Liability_Limitation_Act", "YouTube_copyright_strike", "Copyright_infringement", "Viacom_(2005%E2%80%93present)", "Mediaset", "Premier_League", "Lenz_v._Universal_Music_Corp.", "Fair_use", "Smith_v._Summit_Entertainment_LLC", "Causes_of_action", "Universal_Music_Group", "Content_ID_(system)", "Viacom_(2005%E2%80%93present)", "YouTube_copyright_issues", "Criticism_of_Google#YouTube", "Censorship_by_Google#YouTube", "Shock_site", "Recommender_system", "Elsagate", "Pedophilia", "Post-traumatic_stress_disorder", "National_Center_for_Missing_and_Exploited_Children", "Holocaust_denial", "Hillsborough_disaster", "Anthony_Weiner", "Anwar_al-Awlaki", "Fake_news", "Encyclop%C3%A6dia_Britannica", "Notre-Dame_de_Paris_fire", "September_11_attacks", "LGBTQ%2B", "2017_Las_Vegas_shooting", "Anthony_Bourdain", "Alex_Jones", "Media_Matters", "Zeynep_Tufekci", "Modern_flat_Earth_societies", "9/11_Truth_movement", "Tor_Browser", "Scientific_consensus_on_climate_change", "2022_Philippine_general_election", "Supremacism", "Hate_speech", "Christchurch_mosque_attacks", "The_New_York_Times", "ISIS", "Online_hate_speech", "Stefan_Molyneux", "David_Duke", "Richard_B._Spencer", "Reddit", "R/The_Donald", "Misinformation_related_to_the_COVID-19_pandemic", "5G", "Coronavirus_disease_2019", "World_Health_Organization", "Robert_F._Kennedy_Jr.", "Joseph_Mercola", "RT_Deutsch", "RT_(TV_network)", "Climate_change_denial", "FamilyOFive", "Fantastic_Adventures_scandal", "Elsagate", "Child_protection", "FamilyOFive", "Child_abuse", "Fantastic_Adventures_scandal", "YouTube_Kids", "Elsagate", "Toy_Freaks", "Cocomelon_%E2%80%93_Nursery_Rhymes", "Bloomberg_News", "Campaign_for_a_Commercial-Free_Childhood", "ChuChu_TV", "Consumption_of_Tide_Pods", "Elsagate", "Pedophilia", "Dark_web", "The_Times", "National_Center_for_Missing_and_Exploited_Children", "AT%26T", "The_Walt_Disney_Company", "Dr._Oetker", "Epic_Games", "Nestl%C3%A9", "Child_pornography", "Pok%C3%A9mon_Go", "Club_Penguin", "PewDiePie_vs_T-Series", "Josh_Hawley", "YouTube_Kids", "April_Fools%27_Day", "Rick_Astley", "Never_Gonna_Give_You_Up", "Rickrolling", "ASCII_art", "Keyboard_Cat", "The_Onion", "Internet_celebrity", "Antoine_Dodson", "Sandstorm_(instrumental)", "Darude", "Snoop_Dogg", "Android_(operating_system)", "Emerging_market", "Bluetooth", "Display_resolution", "India", "Nigeria", "Indonesia", "Thailand", "Malaysia", "Vietnam", "Philippines", "Kenya", "South_Africa", "Brazil", "Mexico", "Turkey", "Iraq", "YouTube_Kids", "Google", "Android_(operating_system)", "IOS", "Mobile_app", "LG_Electronics", "Samsung_Electronics", "Sony", "Smart_TV", "Android_TV", "Apple_TV", "YouTube_Music", "Lyor_Cohen", "300_Entertainment", "Warner_Music_Group", "Spotify", "Apple_Music", "YouTube_Premium", "YouTube_Originals", "Comparison_of_on-demand_music_streaming_services", "List_of_YouTube_original_programming", "TikTok", "Instagram_Stories", "Snapchat", "YouTube_TV", "Over-the-top_content", "Multichannel_video_programming_distributor", "New_York_City", "Los_Angeles", "Chicago", "Philadelphia", "San_Francisco", "American_Broadcasting_Company", "CBS", "The_CW", "Fox_Broadcasting_Company", "NBC", "The_Walt_Disney_Company", "CBS_Corporation", "21st_Century_Fox", "NBCUniversal", "Turner_Broadcasting_System", "Bravo_(U.S._TV_network)", "USA_Network", "Syfy", "Disney_Channel", "CNN", "Cartoon_Network", "E!", "Fox_Sports_1", "Freeform_(TV_network)", "FX_(TV_network)", "ESPN", "Showtime_(TV_network)", "Fox_Soccer_Plus", "YouTube_Premium", "Social_impact_of_YouTube", "Old_media", "Complete_Music_Update", "Forbes", "Billboard_(magazine)", "Billboard_Hot_100", "Peabody_Awards", "TED_(conference)", "Chris_Anderson_(entrepreneur)", "Johannes_Gutenberg", "Khan_Academy", "Michael_Noer_(editor)", "Disruptive_innovation", "George_Foster_Peabody_Award", "Speakers%27_Corner", "Pew_Research_Center", "CNN/YouTube_presidential_debates", "TechPresident", "Arab_Spring", "Philip_N._Howard", "Joseph_Kony", "Kony_2012", "Lindsey_Graham", "White_House", "Iman_Crosson", "Patient_Protection_and_Affordable_Care_Act", "New_media", "Innocence_of_Muslims", "Reactions_to_Innocence_of_Muslims", "Virginia_Heffernan", "Conservative_talk_radio", "Far-right_politics", "Recommender_system", "Alex_Jones", "Fox_News", "The_Legion_of_Extraordinary_Dancers", "YouTube_Symphony_Orchestra", "We_Are_the_World_25_for_Haiti_(YouTube_edition)", "It_Gets_Better_Project", "Suicide_among_LGBT_youth", "Barack_Obama", "Suicide_of_Amanda_Todd", "Metropolitan_Police", "UK_drill", "Gang#Gang_violence", "Materiality_(auditing)", "Netflix", "Amazon_Prime_Video", "NBC", "BBC", "Metro-Goldwyn-Mayer", "Lions_Gate_Entertainment", "CBS", "Fox_Broadcasting_Company", "Walt_Disney_Studios_Motion_Pictures", "The_Guardian", "Offensive_content", "H3h3Productions", "Wall_Street_Journal", "Johnny_Rebel_(singer)", "Chief_Keef", "AdSense", "TubeMogul", "YouTube_Play_Button", "Censorship_by_Google#Advertiser-friendly_content", "Phillip_DeFranco", "Vlogbrothers", "Boing_Boing", "Nintendo", "Censorship_of_YouTube", "North_Korea", "Internet_censorship_in_China", "Morocco", "Turkey", "Libya", "Pakistan", "Internet_censorship_in_Iran", "Blocking_of_YouTube_videos_in_Germany", "YouTube_Kids", "Censorship_of_YouTube#Censorship_of_LGBT_content_in_Restricted_Mode", "China", "North_Korea", "Iran", "Syria", "Turkmenistan", "Uzbekistan", "Tajikistan", "Eritrea", "Sudan", "South_Sudan", "Chinese_Communist_Party", "50_Cent_Party", "State-sponsored_Internet_propaganda", "Thailand", "Monarchy_of_Thailand", "Political_status_of_Western_Sahara", "Mustafa_Kemal_Atat%C3%BCrk", "Deniz_Baykal", "Jyllands-Posten_Muhammad_cartoons_controversy", "Muhammad", "Virtual_private_network", "Everybody_Draw_Mohammed_Day", "Innocence_of_Muslims", "Benghazi", "Abu_Salim_prison", "Muammar_Gaddafi", "Human_Rights_Watch", "2011_Libyan_Civil_War", "Afghanistan", "Bangladesh", "Pakistan", "Sudan", "Reactions_to_Innocence_of_Muslims", "Innocence_of_Muslims", "Chechnya", "Egypt", "Portal:Companies", "Portal:Internet", "Portal:San_Francisco_Bay_Area", "Portal:United_States", "Alternative_media", "BookTube", "BreadTube", "CNN-YouTube_presidential_debates", "Comparison_of_video_hosting_services", "List_of_Internet_phenomena", "List_of_most_viewed_online_videos_in_the_first_24_hours", "List_of_online_video_platforms", "List_of_most-disliked_YouTube_videos", "List_of_most-liked_YouTube_videos", "List_of_most-viewed_YouTube_videos", "List_of_most-subscribed_YouTube_channels", "List_of_YouTubers", "Ouellette_v._Viacom_International_Inc.", "Reply_Girls", "Viacom_International_Inc._v._YouTube,_Inc.", "YouTube_Awards", "YouTube_Creator_Awards", "YouTube_Instant", "YouTube_Live", "YouTube_Multi_Channel_Network", "YouTube_Music_Awards", "YouTube_Rewind", "YouTube_Theater", "GitHub", "TubeFilter", "USA_Today", "Wayback_Machine", "The_Washington_Post", "The_Daily_Telegraph", "Wayback_Machine", "Wayback_Machine", "The_Daily_Telegraph", "Business_Insider", "USA_Today", "Fortune_(magazine)", "Gizmodo", "The_New_York_Times", "Variety_(magazine)", "The_Hollywood_Reporter", "USA_Today", "CNNMoney", "CNN", "TechCrunch", "AOL", "TechCrunch", "AOL", "The_Age", "The_Daily_Telegraph", "ComScore", "CNET", "CBS_Interactive", "Thomson_Reuters", "The_Daily_Telegraph", "Mashable", "The_Verge", "Vox_Media", "CNET", "CBS_Interactive", "Mashable", "BBC_News", "San_Jose_Mercury_News", "TechCrunch", "AOL", "Engadget", "AOL", "ISSN_(identifier)", "CNN", "CNN_Business", "Doi_(identifier)", "ISBN_(identifier)", "S2CID_(identifier)", "GitHub", "Ars_Technica", "CNET", "CBS_Interactive", "University_of_Washington", "University_of_Washington", "MSN", "CNET", "CBS_Interactive", "CNET", "CBS_Interactive", "The_Guardian", "TechCrunch", "AOL", "ISSN_(identifier)", "The_Verge", "The_Verge", "Ars_Technica", "TechCrunch", "AOL", "The_Daily_Telegraph", "The_Huffington_Post", "Wayback_Machine", "CNET", "CBS_Interactive", "CNET", "CBS_Interactive", "The_Washington_Post", "Apple_Inc.", "TechCrunch", "AOL", "Business_Insider", "Gizmodo", "Variety_(magazine)", "El_Universal_(Mexico_City)", "Mashable", "Mashable", "Mashable", "PC_World", "The_Hindu", "Wayback_Machine", "Wayback_Machine", "Wayback_Machine", "Wayback_Machine", "Wayback_Machine", "TONO", "Swissinfo", "Wiener_Zeitung", "Dnevnik_(Slovenia)", "The_Nation_(Thailand)", "The_Himalayan_Times", "The_Express_Tribune", "CNET", "CBS_Interactive", "The_Daily_Telegraph", "The_Daily_Dot", "Complex_Media,_Inc.", "Dan_Sabbagh", "The_Guardian", "Guardian_News_and_Media_Limited", "Brady_Haran", "Numberphile", "Time_(magazine)", "CNNMoney.com", "The_Daily_Telegraph", "BBC_News", "BBC_News", "ISSN_(identifier)", "ISSN_(identifier)", "ISSN_(identifier)", "The_New_York_Times", "The_Daily_Telegraph", "The_Verge", "Bloomberg_L.P.", "The_Washington_Post", "ISSN_(identifier)", "ISSN_(identifier)", "ISSN_(identifier)", "CNBC", "Doi_(identifier)", "ISSN_(identifier)", "ISSN_(identifier)", "Rappler", "Manila", "The_Verge", "The_Washington_Post", "The_New_York_Times", "CNBC", "The_Guardian", "The_Verge", "The_Verge", "The_Guardian", "The_Verge", "Time_(magazine)", "The_Verge", "The_Washington_Post", "The_Guardian", "ISSN_(identifier)", "New_York_(magazine)", "Time_(magazine)", "The_New_York_Times", "Forbes_(magazine)", "Variety_(magazine)", "The_Verge", "Newshub", "ABC_News", "BuzzFeed", "Wired_(magazine)", "Bloomberg_News", "Vice_Media", "The_Verge", "New_Statesman", "Variety_(magazine)", "ISSN_(identifier)", "Bloomberg_News", "The_Verge", "CNBC", "Engadget", "The_Verge", "BBC_News", "The_Verge", "The_New_York_Times", "Engadget", "Wall_Street_Journal", "TechCrunch", "AOL", "The_Daily_Telegraph", "Mashable", "Ad_Age", "Engadget", "AOL", "The_Verge", "Vox_Media", "Variety_(magazine)", "Variety_(magazine)", "Ars_Technica", "The_Verge", "Vox_Media", "Business_Insider", "Los_Angeles_Times", "Gizmodo", "Next_Big_Sound", "TED_(conference)", "Wayback_Machine", "The_Washington_Post", "Pew_Research_Center", "Pew_Research_Center", "Whitehouse.gov", "NARA", "Pew_Research_Center", "TED_(conference)", "The_Verge", "Alphabet_Inc.", "Engadget", "BBC_News", "The_Washington_Post", "Jeff_Bezos", "TechCrunch", "AOL", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "The_Guardian", "Variety_(magazine)", "The_Verge", "Vox_Media", "USA_Today", "Business_Insider", "CNET", "CBS_Interactive", "Boing_Boing", "IGN", "J2_Global", "Forbes", "Recode", "Vox_Media", "Kotaku", "Univision_Communications", "David_Meerman_Scott", "Reporters_Without_Borders", "BBC_News", "Alexa_Internet", "Business_Insider", "BBC", "NBC_News", "ABC_News_(Australia)", "Wayback_Machine", "Wayback_Machine", "The_Washington_Post", "Bloomberg_L.P.", "Business_Insider", "Brady_Haran", "ISBN_(identifier)", "ISBN_(identifier)", "The_New_York_Times", "PBS", "Off_Book_(web_series)", "Template_talk:YouTube_navbox", "Google", "History_of_YouTube", "Social_impact_of_YouTube", "Susan_Wojcicki", "Lyor_Cohen", "Susanne_Daniels", "John_Farrell_(businessman)", "Robert_Kyncl", "Steve_Chen", "Chad_Hurley", "Salar_Kamangar", "Jawed_Karim", "Udi_Manber", "Yu_Pan", "List_of_most-subscribed_YouTube_channels", "List_of_most-subscribed_YouTube_Music_artists", "List_of_most-viewed_YouTube_channels", "Multi-channel_network", "List_of_multi-channel_networks", "YouTube_(channel)", "YouTube_Original_Channel_Initiative", "BookTube", "BreadTube", "Charlie_Bit_My_Finger", "YouTube_Creator_Awards", "Draw_My_Life", "Let%27s_Play", "Me_at_the_zoo", "PewDiePie_vs_T-Series", "Reply_girl", "Viral_video", "YouTube_Poop", "YouTube_Rewind", "YouTube_Rewind_2018:_Everyone_Controls_Rewind", "YouTube_Rewind_2019:_For_the_Record", "YouTuber", "List_of_YouTubers", "Virtual_YouTuber", "List_of_most-viewed_YouTube_videos", "List_of_most-viewed_online_videos_in_the_first_24_hours", "List_of_most-viewed_Arabic_music_videos_on_YouTube", "List_of_most-viewed_French_music_videos_on_YouTube", "List_of_most-viewed_Indian_YouTube_videos", "List_of_most-viewed_Pakistani_YouTube_videos", "List_of_most-liked_YouTube_videos", "List_of_most-disliked_YouTube_videos", "Category:2000s_YouTube_videos", "Category:2010s_YouTube_videos", "Category:2020s_YouTube_videos", "Criticism_of_YouTube", "2020_Korean_YouTube_backdoor_advertising_controversy", "Censorship_by_YouTube", "YouTube_copyright_issues", "Content_ID_(system)", "YouTube_copyright_strike", "Elsagate", "Fantastic_Adventures_scandal", "Garcia_v._Google,_Inc.", "2012_Kohistan_video_case", "Lenz_v._Universal_Music_Corp.", "YouTube_and_privacy", "Reactions_to_Innocence_of_Muslims", "2011_Slovenian_YouTube_incident", "Viacom_International_Inc._v._YouTube,_Inc.", "YouTube_Kids", "YouTube_Music", "YouTube_Premium", "List_of_YouTube_Premium_original_programming", "YouTube_TV", "YouTube_Awards", "CNN/YouTube_presidential_debates", "YouTube_Comedy_Week", "KSI_vs._Logan_Paul", "KSI_vs._Logan_Paul_II", "YouTube_Live", "YouTube_Music_Awards", "2013_YouTube_Music_Awards", "2015_YouTube_Music_Awards", "Playlist_Live", "Project_for_Awesome", "Social_in_the_City", "YouTube_Space_Lab", "YouTube_Symphony_Orchestra", "Internet_Symphony_No._1", "Mothership_(composition)", "Tubecon", "VidCon", "YouTube_API", "Censorship_of_YouTube", "Blocking_of_YouTube_videos_in_Germany", "Google_Preferred", "Vevo", "YouTube_headquarters_shooting", "YouTube_in_education", "YouTube_Instant", "YouTube_Space", "YouTube_suspensions", "YouTube_Theater", "Template_talk:Google_LLC", "Google", "Alphabet_Inc.", "History_of_Google", "List_of_Android_apps_by_Google", "List_of_Google_Easter_eggs", "List_of_Google_April_Fools%27_Day_jokes", "List_of_mergers_and_acquisitions_by_Alphabet", "Google_Ads", "Google_AI", "Google_Brain", "Android_(operating_system)", "Booting_process_of_Android_devices", "Android_recovery_mode", "Android_software_development", "Android_version_history", "Google_China", "Google_Chrome", "Google_Cloud_Platform", "Google_Glass", "Google.org", "Google_Crisis_Response", "Google_Public_Alerts", "RechargeIT", "Google_Health", "Google_Maps", "Google_Nest", "Google_Pixel", "Google_Search", "Timeline_of_Google_Search", "Sidewalk_Labs", "Sidewalk_Toronto", "Google_Stadia", "History_of_YouTube", "Me_at_the_zoo", "Social_impact_of_YouTube", "YouTuber", "Category:Google_employees", "Krishna_Bharat", "Vint_Cerf", "Jeff_Dean", "John_Doerr", "Sanjay_Ghemawat", "Al_Gore", "John_L._Hennessy", "Urs_H%C3%B6lzle", "Salar_Kamangar", "Ray_Kurzweil", "Ann_Mather", "Alan_Mulally", "Sundar_Pichai", "Ruth_Porat", "Rajen_Sheth", "Hal_Varian", "Susan_Wojcicki", "Andy_Bechtolsheim", "Sergey_Brin", "David_Cheriton", "Matt_Cutts", "David_Drummond_(businessman)", "Alan_Eustace", "Timnit_Gebru", "Omid_Kordestani", "Paul_Otellini", "Larry_Page", "Patrick_Pichette", "Eric_Schmidt", "Ram_Shriram", "Amit_Singhal", "Shirley_M._Tilghman", "Rachel_Whetstone", "111_Eighth_Avenue", "Androidland", "Google_barges", "Chelsea_Market", "Chrome_Zone", "Google_data_centers", "Google_Modular_Data_Center", "Googleplex", "St._John%27s_Terminal", "YouTube_Space", "YouTube_Theater", "Croscore_fonts", "Product_Sans#Variations", "Noto_fonts", "Product_Sans", "Roboto", "Google_logo", "Google_Doodle", "Doodle_Champion_Island_Games", "Magic_Cat_Academy", "Google_logo#Favicon", "Material_Design", "Comparison_of_Material_Design_implementations", "Android_Developer_Challenge", "Android_Developer_Day", "Android_Developer_Lab", "Google_Code-in", "Google_Code_Jam", "Google_Developer_Day", "Google_Developers_Live", "Doodle4Google", "Google_I/O", "Jigsaw_(company)", "Living_Stories", "Google_Lunar_X_Prize", "Google_Mapathon", "Google_Science_Fair", "Google_Summer_of_Code", "Talks_at_Google", "YouTube_Awards", "CNN/YouTube_presidential_debates", "YouTube_Comedy_Week", "YouTube_Live", "YouTube_Music_Awards", "YouTube_Space_Lab", "YouTube_Symphony_Orchestra", "20%25_Project", "A_Google_A_Day", "Area_120", "Reply_(Google)", "Tables_(Google)", "Google_ATAP", "Google_Data_Liberation_Front", "Data_Transfer_Project", "Digital_News_Initiative", "Digital_Unlocked", "Dragonfly_(search_engine)", "Google_Founders%27_Award", "Google_Free_Zone", "Google_Get_Your_Business_Online", "Google_for_Education", "Google_for_Startups", "Google_Labs", "Liquid_Galaxy", "Made_with_Code", "Google_Maori", "Google_News_Lab", "Google_PowerMeter", "Project_Nightingale", "Privacy_Sandbox", "RechargeIT", "Project_Shield", "Project_Starline", "Google_Student_Ambassador_Program", "Dunant_(submarine_communications_cable)", "Grace_Hopper_(submarine_communications_cable)", "Project_Sunroof", "Where_on_Google_Earth_is_Carmen_Sandiego%3F", "YouTube_Creator_Awards", "YouTube_Next_Lab_and_Audience_Development_Group", "YouTube_Original_Channel_Initiative", "Project_Zero", "Criticism_of_Google", "2018_Google_data_breach", "2018_Google_walkouts", "2020_Google_services_outages", "Alphabet_Workers_Union", "Censorship_by_Google", "DeGoogle", "Dragonfly_(search_engine)", "FairSearch", "Google%27s_Ideological_Echo_Chamber", "Is_Google_Making_Us_Stupid%3F", "Google_litigation", "Predictions_of_the_end_of_Google", "Privacy_concerns_regarding_Google", "Google_Street_View_privacy_concerns", "San_Francisco_tech_bus_protests", "Google_worker_organization", "YouTuber_back_advertisement_controversy_in_Korea_in_2020", "Censorship_by_Google#YouTube", "YouTube_copyright_issues", "YouTube_copyright_strike", "Elsagate", "Fantastic_Adventures_scandal", "YouTube_headquarters_shooting", "2012_Kohistan_video_case", "Reactions_to_Innocence_of_Muslims", "2011_Slovenian_YouTube_incident", "Google_Developers", "Android_(operating_system)", "Android_Automotive", "Glass_OS", "Android_Go", "GLinux", "Goobuntu", "Android_Things", "Android_TV", "Wear_OS", "Chrome_OS", "Chromium_OS", "Neverware", "Fuchsia_(operating_system)", "Google_TV_(operating_system)", "Accelerated_Mobile_Pages", "Angular_(web_framework)", "AngularJS", "ARCore", "Google_APIs", "Blockly", "Google_Chart_API", "Google_Charts", "Dialogflow", "Exposure_Notification", "Fast_Pair", "Federated_Learning_of_Cohorts", "Google_File_System", "FlatBuffers", "Flutter_(software)", "Gears_(software)", "GRPC", "Gson", "Google_Guava", "Google_Guice", "Guetzli", "GVisor", "MapReduce", "Google_mobile_services", "Google_Neural_Machine_Translation", "OpenSocial", "Google_Pack", "Polymer_(library)", "Protocol_Buffers", "Reqwireless", "Google_Shell", "Skia_Graphics_Engine", "Tango_(platform)", "TensorFlow", "Google_Test", "WaveNet", "Weave_(protocol)", "Google_Web_Accelerator", "WebRTC", "Google_App_Engine", "Google_Apps_Script", "Google_Cloud_Platform", "Anvato", "Bigtable", "Chronicle_Security", "VirusTotal", "Google_Compute_Engine", "Google_Cloud_Connect", "Google_Cloud_Dataflow", "Google_Cloud_Datastore", "Kaggle", "Looker_(company)", "Google_Cloud_Messaging", "Google_Cloud_Shell", "Stackdriver", "Google_Cloud_Storage", "Firebase", "Firebase_Cloud_Messaging", "Crashlytics", "Global_IP_Solutions", "Internet_Low_Bitrate_Codec", "Internet_Speech_Audio_Codec", "Gridcentric,_Inc.", "ITA_Software", "Kubernetes", "LevelDB", "Orbitera", "SageTV", "Android_Cloud_to_Device_Messaging", "Android_Studio", "App_Inventor_for_Android", "Google_App_Maker", "Google_App_Runtime_for_Chrome", "AppSheet", "Bazel_(software)", "Google_Chrome_Frame", "Google_Closure_Tools", "Cpplint", "Google_Gadgets", "GData", "Gerrit_(software)", "GYP_(software)", "Google_Kythe", "Google_Lighthouse", "Google_Mashup_Editor", "Google_Native_Client", "Google_Optimize", "OpenRefine", "OR-Tools", "Google_PageSpeed_Tools", "Google_Plugin_for_Eclipse", "Google_Programmable_Search_Engine", "Google_Public_DNS", "ReCAPTCHA", "Google_Search_Console", "Sitemaps", "Google_Swiffy", "Tesseract_(software)", "Trendalyzer", "Google_Wave_Federation_Protocol", "Google_Web_Toolkit", "Google_Hummingbird", "PageRank", "Google_matrix", "Google_Panda", "Google_Penguin", "Google_Pigeon", "RankBrain", "BigQuery", "Google_Chrome_Experiments", "Android_App_Bundle", "Android_application_package", "AV1", "On2_Technologies", "VP3", "VP6", "VP8", "Libvpx", "VP9", "WebM", "WebP", "Web_Open_Font_Format", "Open_Location_Code", "Flutter_(American_company)", "Caja_project", "Dart_(programming_language)", "Go_(programming_language)", "Sawzall_(programming_language)", "Viewdle", "Google_Web_Server", "List_of_Google_products", "Google_Currents_(2011%E2%80%932013)", "Green_Throttle_Games", "Owlchemy_Labs", "Oyster_(company)", "PaperofRecord.com", "Google_Play", "Google_Play_Books", "Google_Play_Games", "List_of_most-downloaded_Google_Play_applications", "Google_Play_Music", "Google_Play_Newsstand", "Google_Play_Pass", "Google_Play_Services", "Google_Podcasts", "Quick,_Draw!", "Google_Santa_Tracker", "Google_Stadia", "List_of_Stadia_games", "Typhoon_Studios", "Google_TV_(service)", "Vevo", "Google_Video", "YouTube_API", "BandPage", "Content_ID_(system)", "FameBit", "YouTube_Instant", "YouTube_Kids", "YouTube_Music", "YouTube_(channel)", "Google_Preferred", "YouTube_Premium", "List_of_YouTube_Premium_original_programming", "YouTube_Rewind", "RightsFlow", "YouTube_TV", "Google_Allo", "Bump_(application)", "Google_Buzz", "Google_Chat", "Google_Contacts", "Google_Currents", "Dodgeball_(service)", "Google_Duo", "Google_Fi", "Google_Friend_Connect", "Gizmo5", "Google%2B", "Gmail", "History_of_Gmail", "Inbox_by_Gmail", "Gmail_interface", "Google_Groups", "Google_Hangouts", "Google_Helpouts", "Google_IME", "Google_Japanese_Input", "Google_Pinyin", "Google_transliteration", "Jaiku", "Meebo", "Google_Meet", "Messages_(Google)", "Google_Moderator", "Neotonic_Software", "Orkut", "Postini", "Google_Schemer", "Google_Spaces", "Sparrow_(email_client)", "Google_Talk", "Google_Translate", "Google_Translator_Toolkit", "Google_Voice", "GOOG-411", "Google_Wave", "Word_Lens", "Quest_Visual", "Google_Search", "Aardvark_(search_engine)", "Akwan_Information_Technologies", "Google_Alerts", "Google_Answers", "Google_Base", "Google_Blog_Search", "Google_Books", "Google_Ngram_Viewer", "Google_Code_Search", "Google_Dataset_Search", "Google_Dictionary", "Google_Directory", "Google_Fast_Flip", "Google_Flu_Trends", "Google_Finance", "Google_Goggles", "Googlebot", "Google_Images", "Google_Image_Labeler", "Google_Image_Swirl", "Kaltix", "Google_Knowledge_Graph", "Freebase_(database)", "Metaweb", "Like.com", "Google_News", "Google_News_Archive", "Google_News_%26_Weather", "Google_Patents", "People_Cards", "Google_Personalized_Search", "Google_Public_Data_Explorer", "Google_Questions_and_Answers", "SafeSearch", "Google_Scholar", "Google_SearchWiki", "Google_Shopping", "Google_Catalogs", "Google_Express", "Google_Squared", "Tenor_(website)", "Google_Travel", "Google_Flights", "Google_Trends", "Google_Insights_for_Search", "Google_Voice_Search", "WDYL_(search_engine)", "Google_Earth", "ImageAmerica", "Google_Maps", "Google_Latitude", "Google_Map_Maker", "Google_Maps_Navigation", "Google_Maps_pin", "Google_Street_View", "Coverage_of_Google_Street_View", "Street_View_Trusted", "Waze", "Google_Ad_Manager", "AdMob", "Google_Ads", "Adscape", "Google_AdSense", "Google_Attribution", "BebaPay", "Google_Business_Groups", "Google_Checkout", "Google_Contributor", "DMarc_Broadcasting", "DoubleClick", "Google_Affiliate_Network", "Invite_Media", "Google_Analytics", "Google_Analytics", "Google_Data_Studio", "Urchin_(software)", "Google_Pay", "Google_Pay_Send", "Tez_(software)", "PostRank", "Google_Primer", "Softcard", "Wildfire_Interactive", "Google_Bookmarks", "Google_Browser_Sync", "Google_Calendar", "Google_Cloud_Search", "Google_Desktop", "Google_Docs_Editors", "Google_Docs", "Google_Drawings", "Google_Forms", "Google_Fusion_Tables", "Google_Keep", "Google_Sheets", "Google_Slides", "Google_Sites", "Google_Drive", "Fflick", "Files_(Google)", "IGoogle", "Jamboard", "Google_Notebook", "Google_One", "Google_Photos", "Apture", "Blogger_(service)", "Pyra_Labs", "Google_Domains", "FeedBurner", "Google_One_Pass", "Google_Page_Creator", "Google_Sites", "Google_Web_Designer", "Quickoffice", "Google_Quick_Search_Box", "Google_Surveys", "Google_Sync", "Google_Calendar", "Google_Toolbar", "Google_Account", "Google_Dashboard", "Google_Takeout", "Android_Auto", "Android_Beam", "Google_Arts_%26_Culture", "Google_Assistant", "Google_Authenticator", "ZygoteBody", "BufferBox", "Google_Building_Maker", "Google_Cast", "List_of_apps_with_Google_Cast_support", "Google_Chrome", "Google_Chrome_App", "Chromium_(web_browser)", "Dinosaur_Game", "GreenBorder", "Chrome_Remote_Desktop", "Google_Chrome_version_history", "Chrome_Web_Store", "Google_Classroom", "Google_Cloud_Print", "Crowdsource_(app)", "Google_Expeditions", "Google_Family_Link", "Find_My_Device", "Google_Fit", "Google_Fonts", "Gboard", "Google_Gesture_Search", "Google_Camera", "Google_Lens", "Snapseed", "Nik_Software", "Panoramio", "Google_Photos", "Picasa", "Picasa_Web_Albums", "Picnik", "Impermium", "Knol", "Google_Lively", "Live_Transcribe", "MyTracks", "Nearby_Share", "Google_Now", "Google_Offers", "Google_Opinion_Rewards", "Google_Person_Finder", "PlinkArt", "Poly_(website)", "Google_Question_Hub", "Read_Along", "Google_Reader", "Google_Safe_Browsing", "Google_Sidewiki", "SlickLogin", "Socratic_(Google)", "Sound_Amplifier", "Speech_Services", "Google_Station", "Google_Store", "Google_TalkBack", "Tilt_Brush", "Google_URL_Shortener", "Wavii", "Google_Web_Light", "Google_WiFi", "Google_Workspace", "Google_Workspace_Marketplace", "Android_One", "Google_Nexus", "Nexus_One", "Nexus_S", "Galaxy_Nexus", "Nexus_4", "Nexus_5", "Nexus_6", "Nexus_5X", "Nexus_6P", "Comparison_of_Google_Nexus_smartphones", "Google_Pixel", "Pixel_(1st_generation)", "Pixel_2", "Pixel_3", "Pixel_3a", "Pixel_4", "Pixel_4a", "Pixel_5", "Pixel_5a", "Pixel_6", "Comparison_of_Google_Pixel_smartphones", "List_of_Google_Play_edition_devices", "Project_Ara", "Chromebook", "Google_Nexus", "Nexus_7_(2012)", "Nexus_7_(2013)", "Nexus_10", "Nexus_9", "Comparison_of_Google_Nexus_tablets", "Google_Pixel", "Chromebook_Pixel", "Pixelbook", "Pixelbook_Go", "Pixel_C", "Pixel_Slate", "Chromebit", "Chromebox", "Google_Clips", "Chromecast", "Nexus_Player", "Nexus_Q", "Dropcam", "Fitbit", "List_of_Fitbit_products", "Google_Nest", "Google_Nest#Nest_Cam_Indoor", "Google_Nest#Nest_Doorbell", "Google_Nest#Nest_Protect", "Google_Nest#Nest_Secure", "Google_Nest_(smart_speakers)", "Nest_Thermostat", "Google_Nest_Wifi", "Google_OnHub", "Pixel_Buds", "Google_Pixel#Pixel_Stand", "Pixel_Visual_Core", "Google_Search_Appliance", "Sycamore_processor", "Tensor_Processing_Unit", "Google_Cardboard", "Google_Contact_Lens", "Google_Daydream", "Google_Glass", "Don%27t_be_evil", "Gayglers", "Google_(verb)", "Googlearchy", "Google_bombing", "Google_effect", "Googlefight", "Google_hacking", "Googleshare", "Google_tax", "Googlewhack", "Googlization", "Rooting_(Android)", "YouTube_Poop", "Category:Documentary_films_about_Google", "Google_Behind_the_Screen", "Google_Current", "Google_Maps_Road_Trip", "Google_The_Thinking_Factory", "Google_and_the_World_Brain", "The_Creepy_Line", "Category:Books_about_Google", "Google_Hacks", "The_Google_Story", "Google_Volume_One", "Googled:_The_End_of_the_World_as_We_Know_It", "How_Google_Works", "I%27m_Feeling_Lucky_(book)", "In_the_Plex", ".google", "Google_Developer_Expert", "Google_Feud", "Google_Me_(film)", "Google_Me_(song)", "The_Billion_Dollar_Code", "The_Internship", "List_of_Google_products#Discontinued_products_and_services", "Category:Google", "Outline_of_Google", "Template_talk:Alphabet_Inc.", "Alphabet_Inc.", "Google_Fiber", "Calico_(company)", "CapitalG", "DeepMind", "AlphaFold", "AlphaGo", "AlphaGo_Zero", "AlphaStar_(software)", "AlphaZero", "Master_(software)", "MuZero", "WaveNet", "Google", "DoubleClick", "Firebase", "Fitbit", "Tenor_(website)", "Waze", "GV_(company)", "Isomorphic_Labs", "Verily", "X_Development", "Waymo", "Wing_(company)", "Boston_Dynamics", "Chronicle_Security", "Jigsaw_(company)", "Loon_LLC", "Makani_(company)", "Meka_Robotics", "Google_Nest", "Redwood_Robotics", "Sidewalk_Labs", "Sidewalk_Toronto", "Sundar_Pichai", "Chief_executive_officer", "Ruth_Porat", "Chief_financial_officer", "Larry_Page", "Chief_executive_officer", "Sergey_Brin", "President_(corporate_title)", "David_Drummond_(businessman)", "Chief_legal_officer", "Sergey_Brin", "John_Doerr", "John_L._Hennessy", "Ann_Mather", "Alan_Mulally", "Larry_Page", "Sundar_Pichai", "Ram_Shriram", "Roger_W._Ferguson_Jr.", "Eric_Schmidt", "Diane_Greene", "Andrew_Conrad", "Tony_Fadell", "Arthur_D._Levinson", "David_Krane", "Ruth_Porat", "Astro_Teller", "Category:Alphabet_Inc.", "Portal:Companies", "Portal:Internet", "Template_talk:Video_digital_distribution_platforms", "Online_video_platform", "Digital_library", "Streaming_media", "Video_on_demand", "56.com", "Aparat", "Apple_TV_app", "AcFun", "AfreecaTV", "Bilibili", "BitChute", "Brightcove", "Buzznet", "Dailymotion", "Daum_(web_portal)", "Endeavor_Streaming", "The_Film_Detective", "FilmOn", "Flickr", "Fotki", "Funny_or_Die", "Funshion", "GOG.com", "GyaO", "IMDb#IMDb_TV", "Internet_Archive", "IQIYI", "Le.com", "Mango_TV", "Medici.tv", "Metacafe", "MeWatch", "Niconico", "Noggin_(brand)", "OverDrive_Inc", "Pandora_TV", "PeerTube", "PictureBox_Films", "Pluto_TV", "Pok%C3%A9mon_(TV_series)", "Rumble_(website)", "Rutube", "SchoolTube", "Showroom_(streaming_service)", "Sohu", "TeacherTube", "Teaching_Channel", "Telly_Inc", "ThePlatform", "Trilulilu", "Tudou", "Twitch_(service)", "VBOX7", "Vimeo", "Vudu", "Tencent_Video", "Xunlei", "Youku", "YouTube_Kids", "Zattoo", "Aha_(streaming_service)", "Amazon_Prime_Video", "Fandango_Media", "Vudu", "Google_TV_(service)", "ITunes_Store", "Microsoft_Movies_%26_TV", "Movies_Anywhere", "Nintendo_eShop", "Rakuten_TV", "Adober_Studios", "Azubu", "BBC_Store", "Blip_(website)", "Break.com", "CinemaNow", "Daisuki_(website)", "Disney_Movies_Anywhere", "Fearnet", "FilmStruck", "Flixster", "Global_Wrestling_Network", "Google_Video", "Hitbox_(service)", "Imeem", "IMesh", "Intel_AppUp", "In2TV", "Joost", "Justin.tv", "Kazaa", "LiveLeak", "LoveFilm", "Megaupload", "Mixer_(service)", "MUZU.TV", "MyVideo", "Nintendo_Channel", "Nintendo_Video", "Ovi_(Nokia)", "Openfilm", "Openload", "PlayStation_Video", "7plus", "Presto_(company)", "Putlocker", "Quickflix", "Redbox", "Revver", "Spirit_Show_Network", "Smashcast", "Sony_Connect", "Sony_Entertainment_Network", "Stage6", "Starlight_Networks", "Streamworks_International", "Super_Deluxe", "TalkTalk_TV_Store", "Tank_Top_TV", "TouchVision", "Triton_(content_delivery)", "TroopTube", "Twango", "UltraViolet_(website)", "Impact_Wrestling", "Vdio", "Vessel_(website)", "Viddler", "Vidme", "Vine_(service)", "Vongo_(video_on_demand_service)", "Warner_Archive_Collection", "WeShow", "Windows_Media_Center", "WWE_Classics_on_Demand", "VIAF_(identifier)", "Video_game_live_streaming", "AfreecaTV", "Bilibili", "Caffeine_(service)", "DLive", "Facebook_Gaming", "Steam_(service)", "Twitch_(service)", "YouTube_Gaming", "YY.com", "Mixer_(service)", "Smashcast", "Category:Cloud_gaming"], "Google": ["Google_Search", "Google_(disambiguation)", "Googol", "Googleplex", "Subsidiary", "Limited_liability_company", "Artificial_intelligence", "Advertising", "Cloud_computing", "Computer_software", "Computer_hardware", "Internet", "Menlo_Park,_California", "California", "Larry_Page", "Sergey_Brin", "Googleplex", "Mountain_View,_California", "Sundar_Pichai", "Chief_executive_officer", "Ruth_Porat", "Chief_financial_officer", "Thomas_Kurian", "Chief_executive_office", "Google_Cloud_Platform", "Prabhakar_Raghavan", "Senior_Vice_President", "List_of_Google_products", "Earnings_before_interest_and_taxes", "Net_income", "Asset", "Parent_company", "Alphabet_Inc.", "Eric_Schmidt", "Sergey_Brin", "Larry_Page", "Multinational_corporation", "Technology_company", "Internet", "Online_advertising", "Search_engine", "Cloud_computing", "Big_Tech", "Information_technology", "Amazon_(company)", "Apple_Inc.", "Meta_Platforms", "Facebook", "Microsoft", "Larry_Page", "Sergey_Brin", "Doctor_of_Philosophy", "Stanford_University", "Public_company", "Initial_public_offering", "Alphabet_Inc.", "Holding_company", "Sundar_Pichai", "Alphabet_Workers_Union", "List_of_mergers_and_acquisitions_by_Alphabet", "Search_engine", "Google_Search", "Google_Docs", "Google_Sheets", "Google_Slides", "Gmail", "Google_Calendar", "Google_Drive", "Google_Duo", "Google_Chat", "Google_Meet", "Google_Translate", "Google_Maps", "Waze", "Google_Earth", "Street_View", "Google_Podcasts", "YouTube", "Blogger_(service)", "Google_Keep", "Jamboard", "Google_Photos", "Android_(operating_system)", "Google_Chrome", "Chrome_OS", "Free_and_open-source_software", "Chromium_OS", "Google_Nexus", "Pixel_(smartphone)", "Google_Home", "Google_Wifi_(router)", "Google_Fiber", "Google_Fi", "List_of_most_popular_websites", "YouTube", "Blogger_(service)", "List_of_most_valuable_brands", "Forbes", "Criticism_of_Google", "Privacy_concerns_regarding_Google", "Censorship_by_Google", "Search_neutrality", "Antitrust", "Monopoly", "History_of_Google", "List_of_mergers_and_acquisitions_by_Alphabet", "Alphabet_Inc.", "Larry_Page", "Sergey_Brin", "Larry_Page", "Sergey_Brin", "Stanford_University", "California", "Google_Search", "Robotics", "Willow_Garage", "PageRank", "Relevance_(information_retrieval)", "BackRub_(search_engine)", "Rajeev_Motwani", "Terry_Winograd", "H%C3%A9ctor_Garc%C3%ADa-Molina", "Jeff_Ullman", "RankDex", "Robin_Li", "Baidu", "Googol", "HTML", "Markup_language", "Susan_Wojcicki", "Menlo_Park,_California", "Craig_Silverstein", "Andy_Bechtolsheim", "Sun_Microsystems", "Angel_investor", "Amazon.com", "Jeff_Bezos", "David_Cheriton", "Ram_Shriram", "Menlo_Park,_California", "Venture_capital", "Kleiner_Perkins", "Sequoia_Capital", "Palo_Alto,_California", "Silicon_Valley", "Yahoo!", "Inktomi", "Silicon_Graphics", "Mountain_View,_California", "Googleplex", "Googolplex", "Google_(verb)", "Merriam-Webster_Collegiate_Dictionary", "Oxford_English_Dictionary", "Buffy_the_Vampire_Slayer", "Eric_Schmidt", "Public_company", "Initial_public_offering", "Larry_Page", "Sergey_Brin", "Eric_Schmidt", "Morgan_Stanley", "Credit_Suisse", "Market_capitalization", "Eric_Schmidt", "YouTube", "DoubleClick", "Data_centers", "Motorola_Mobility", "Apple_Inc.", "Microsoft", "Waze", "Google_Maps", "Calico_(company)", "Apple_Inc.", "Arthur_D._Levinson", "DeepMind_Technologies", "London", "Recode", "Apple_Inc.", "Conglomerate_(company)", "Alphabet_Inc.", "Umbrella_company", "Sundar_Pichai", "Chief_executive_officer", "Larry_Page", "Sundar_Pichai", "Prime_Minister_of_India", "Narendra_Modi", "Google%27s_Ideological_Echo_Chamber", "Google_worker_organization", "Dragonfly_(search_engine)", "Project_Maven", "The_New_York_Times", "Andy_Rubin", "Cloud_gaming", "Google_Stadia", "United_States_Department_of_Justice", "Antitrust", "Web_search_engine", "Search_advertising", "PayPal", "Chief_operating_officer", "Google_Pay", "COVID-19_pandemic", "2020_Google_services_outages", "Google_Drive", "YouTube", "Australian_Government", "Ubisoft", "Google_Stadia", "Ubisoft", "The_Wall_Street_Journal", "List_of_Google_products", "Google_Search", "Google_Images", "Search_engine_indexing", "Operator_(computer_programming)", "ComScore", "Google_Search", "Market_share", "Google_Photos", "Google_News", "Google_Books", "AdMob", "AdSense", "AdSense", "DoubleClick", "DoubleClick", "AdSense_for_Mobile", "Google_Analytics", "Google_Ads", "AdSense", "Click_fraud", "Google_Search_Console", "Webmasters", "Gmail", "Email", "Google_Calendar", "Google_Maps", "Satellite_imagery", "Google_Drive", "File_hosting_service", "Google_Docs", "Google_Sheets", "Google_Slides", "Google_Photos", "Google_Keep", "Note-taking", "Google_Translate", "YouTube", "Google_My_Business", "Google_Duo", "Cloud_gaming", "Google_Stadia", "Job_board", "Google_Earth", "Mobile_operating_system", "Wear_OS", "Android_TV", "Android_Auto", "Internet_of_things", "Android_Things", "Google_Chrome", "Chrome_OS", "Nexus_One", "Google_Nexus", "Google_Pixel", "Chromebook", "Chrome_OS", "Chromecast", "Google_Cardboard", "Virtual_reality", "Google_Nest_(smart_speakers)", "Google_Home", "Nest_Audio", "Google_Home_Mini", "Google_Nest_(smart_speakers)#Nest_Mini_(second_generation)", "Google_Home_Max", "Google_Home_Hub", "Nest_Hub_Max", "Google_Nest_Wifi", "Wi-Fi", "Google_Workspace", "Google_Cloud_Platform", "Google_Workspace", "Google_Drive", "Google_Docs", "Google_Sheets", "Google_Slides", "Google_for_Entrepreneurs", "Business_incubator", "Coworking", "Berlin", "London", "Madrid", "Seoul", "S%C3%A3o_Paulo", "Tel_Aviv", "Warsaw", "BigQuery", "Customer_journey", "The_Wall_Street_Journal", "Adobe_Systems", "Oracle_Corporation", "Salesforce.com", "IBM", "Google_Fiber", "Google_Fiber", "Project_Fi", "Initial_public_offering", "Market_capitalization", "Online_advertising", "Mutual_fund", "Class_C_share", "Class_A_share", "NASDAQ", "Ticker_symbol", "Frankfurt_Stock_Exchange", "Corporation_tax_in_the_Republic_of_Ireland#Multinational_tax_schemes", "Google_tax", "Tax_avoidance", "List_of_the_largest_information_technology_companies", "Ireland", "Netherlands", "Bermuda", "Transfer_pricing", "Matt_Brittin", "Public_Accounts_Committee_(United_Kingdom)", "UK_House_of_Commons", "Lobbying_in_the_United_States", "History_of_Google#Name", "Google_(verb)", "Google_logo", "Google_Doodle", "List_of_Google_April_Fools%27_Day_jokes", "List_of_Google_Easter_eggs", "Googol", "PageRank", "Google_(verb)", "Merriam-Webster_Collegiate_Dictionary", "Oxford_English_Dictionary", "Mission_statement", "Don%27t_be_evil", "Google_logo", "Google_Doodle", "Burning_Man", "Larry_Page", "Sergey_Brin", "Intern", "Dennis_Hwang", "Bastille_Day", "April_Fools%27_Day", "Google%27s_hoaxes#2000", "TiSP", "Optical_fiber", "Easter_egg_(media)", "Swedish_Chef", "Pig_Latin", "Leet", "Elmer_Fudd", "International_Talk_Like_a_Pirate_Day", "Klingon_language", "Anagram", "Pride_in_London", "Fortune_(magazine)", "Innovation_Time_Off", "Google_News", "Orkut", "AdSense", "Marissa_Mayer", "The_New_York_Times", "Sexism", "Ageism", "High-Tech_Employee_Antitrust_Litigation", "Silicon_Valley", "UNI_Global_Union", "Googleplex", "Toronto", "Mountain_View,_California", "Googleplex", "Googolplex", "111_Eighth_Avenue", "Manhattan", "New_York_City", "Chelsea_Market", "Hudson_Square", "Ann_Arbor,_Michigan", "Carnegie_Mellon", "Pittsburgh", "Smartphone_applications", "Atlanta,_Georgia", "Austin,_Texas", "Boulder,_Colorado", "Cambridge,_Massachusetts", "San_Francisco", "California", "Seattle,_Washington", "Kirkland,_Washington", "Birmingham,_Michigan", "Reston,_Virginia", "Washington,_D.C.", "Madison,_Wisconsin", "Sydney", "Google_Maps", "London", "London", "Camden_London_Borough_Council", "Hyderabad", "Canada", "United_States", "California", "Googleplex", "Latin_America", "United_Kingdom", "Asia-Pacific", "India", "China", "Africa", "Middle_East", "Michigan", "Georgia_(U.S._state)", "Texas", "Colorado", "Colorado", "Colorado", "Massachusetts", "North_Carolina", "Illinois", "Illinois", "Michigan", "California", "Washington_(state)", "Canada", "California", "Wisconsin", "Florida", "Canada", "California", "New_York_(state)", "Pennsylvania", "California", "Oregon", "California", "Virginia", "California", "California", "California", "Washington_(state)", "California", "Canada", "Washington,_D.C.", "Brazil", "Colombia", "Argentina", "Mexico", "Chile", "Brazil", "Denmark", "Netherlands", "Greece", "Germany", "Belgium", "Denmark", "Republic_of_Ireland", "Germany", "Portugal", "United_Kingdom", "United_Kingdom", "United_Kingdom", "Spain", "Italy", "Russia", "Germany", "Norway", "France", "Czech_Republic", "Sweden", "Austria", "Poland", "Poland", "Switzerland", "India", "Thailand", "China", "China", "India", "Hong_Kong", "India", "Indonesia", "Malaysia", "Australia", "India", "South_Korea", "China", "Singapore", "Australia", "Taiwan", "Japan", "Japan", "United_Arab_Emirates", "Israel", "Turkey", "South_Africa", "Israel", "Google_data_centers", "Google_data_centers", "North_America", "South_America", "Asia", "Europe", "Server_(computing)", "Gartner", "Parallel_computing", "X86", "Submarine_communications_cable", "Chile", "Lisbon", "Portugal", "Lagos", "Nigeria", "Cape_Town", "South_Africa", "New_York_(state)", "Bude", "United_Kingdom", "Bilbao", "Spain", "Solar_panel", "MegaWatt", "Rooftop_photovoltaic_power_station", "Carbon_neutrality", "Nuclear_power_plant", "Google_Energy", "Renewable_energy", "Wind_farm", "North_Dakota", "Federal_Energy_Regulatory_Commission", "100%25_renewable_energy", "UL_(safety_organization)", "Zero_waste#Corporate_initiatives", "Jim_Inhofe", "Climate_change_denial", "State_Policy_Network", "Competitive_Enterprise_Institute", "Google.org", "Google.org", "Climate_change_education", "Plug-in_hybrid", "Electric_vehicle", "Larry_Brilliant", "Mathematical_Sciences_Research_Institute", "Julia_Robinson_Mathematics_Festival", "International_Mathematical_Olympiad", "Legalize_Love", "Gay_rights", "Criticism_of_Google", "Censorship_by_Google", "Privacy_concerns_regarding_Google", "Criticism_of_Google", "San_Francisco_tech_bus_protests", "Google_tax_avoidance", "Criticism_of_Google#Page_rank", "Criticism_of_Google#Copyright_issues", "Google_privacy", "Intellectual_property", "Data_collection", "Internet_privacy", "Energy_consumption", "Monopoly", "Restraint_of_trade", "Anti-competitive_practices", "Patent_infringement", "Internet_censorship_in_the_People%27s_Republic_of_China", "Great_Firewall_of_China", "Censorship_by_Google#China", "Virtual_private_networks", "Proxy_server", "The_Intercept", "Dragonfly_(search_engine)", "PRISM_(surveillance_program)", "Mass_surveillance", "YouTube", "United_States_Department_of_Defense", "Project_Maven", "Drone_strike", "Mozilla", "Firefox", "Google_Chrome", "Reddit", "Subreddit", "DeGoogle", "Grassroots_campaign", "Department_of_Health_and_Human_Services", "Project_Nightingale", "Health_Insurance_Portability_and_Accountability_Act", "St._Louis", "Healthcare", "Ascension_(company)", "Google_litigation", "High-Tech_Employee_Antitrust_Litigation", "Euro", "European_Union", "New_Scientist", "Alphabet,_Inc.", "General_Court_(European_Union)", "European_Commissioner_for_Competition", "General_Court_(European_Union)", "Google%2B", "CNIL", "General_Data_Protection_Regulation", "United_States_Department_of_Justice", "Ken_Paxton", "List_of_advertising_technology_companies", "Facebook", "CNIL", "HTTP_cookie", "Facebook", "Private_browsing", "Google_Chrome", "Reuters", "Sundar_Pichai", "Outline_of_Google", "History_of_Google", "List_of_mergers_and_acquisitions_by_Alphabet", "List_of_Google_products", "Google_China", "Google_logo", "Googlization", "Google.org", "Google_ATAP", "Time_(magazine)", "The_Daily_Telegraph", "The_Independent", "The_Guardian", "InformationWeek", "UBM_plc", "ISSN_(identifier)", "ISSN_(identifier)", "Alexa_Internet", "Forbes", "ISBN_(identifier)", "LCCN_(identifier)", "Vanity_Fair_(magazine)", "Wired_(magazine)", "IEEE_Spectrum", "Business_Insider", "Larry_Page", "Sergey_Brin", "Larry_Page", "Sergey_Brin", "Larry_Page", "CiteSeerX_(identifier)", "Doi_(identifier)", "ISSN_(identifier)", "WebCite", "RankDex", "WHOIS", "San_Francisco_Chronicle", "Wired_(magazine)", "San_Francisco_Chronicle", "Ken_Auletta", "ISBN_(identifier)", "OCLC_(identifier)", "Business_Insider", "Axel_Springer_SE", "CNET", "CBS_Interactive", "ISBN_(identifier)", "CNET", "American_City_Business_Journals", "The_Atlantic", "Business_Insider", "TechCrunch", "AOL", "The_New_York_Times", "TechCrunch", "AOL", "Reuters", "Thomson_Reuters", "Reuters", "The_New_York_Times", "Forbes", "The_Wall_Street_Journal", "IndustryWeek", "Agence_France-Presse", "The_Washington_Post", "Larry_Page", "AppleInsider", "CNET", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "Bloomberg_L.P.", "The_Wall_Street_Journal", "The_New_York_Times", "The_New_York_Times", "Wayback_Machine", "The_New_York_Times", "The_New_York_Times", "Wayback_Machine", "The_New_York_Times", "The_Atlantic", "Bloomberg.com", "The_Verge", "TechCrunch", "The_Wall_Street_Journal", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "InfoWorld", "The_Washington_Post", "The_New_York_Times", "Ars_Technica", "Cond%C3%A9_Nast", "CNET", "Fast_Company_(magazine)", "The_Verge", "Vox_Media", "MacRumors", "The_Verge", "Vox_Media", "TechCrunch", "AOL", "Boy_Genius_Report", "Penske_Media_Corporation", "Time_(magazine)", "Mashable", "Wired_(website)", "Cond%C3%A9_Nast", "Doi_(identifier)", "PMC_(identifier)", "PMID_(identifier)", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "TechCrunch", "AOL", "Ars_Technica", "Cond%C3%A9_Nast", "The_Verge", "Vox_Media", "The_Verge", "Vox_Media", "BBC_News", "BBC", "CNN", "Forbes", "The_Verge", "Vox_Media", "The_Verge", "Vox_Media", "The_Wall_Street_Journal", "TechCrunch", "AOL", "The_Verge", "Vox_Media", "CNN", "CNBC", "The_Washington_Post", "CNN", "CBC_News", "Mashable", "Financial_Times", "The_Washington_Post", "WebCite", "Stanford_University", "Wayback_Machine", "The_Stanford_Daily", "The_Motley_Fool", "MSNBC", "The_Guardian", "Engadget", "AOL", "Andrew_Serwer", "PMID_(identifier)", "Andrew_Serwer", "Andrew_Serwer", "Andrew_Serwer", "The_New_York_Times", "The_New_York_Times", "CNET", "CNET", "CNET", "CBS_Interactive", "The_Wall_Street_Journal", "Dow_Jones_%26_Company", "Wired_(website)", "Cond%C3%A9_Nast", "InformationWeek", "UBM_plc", "WTAE_TV", "Trib_Total_Media", "Quartz_(publication)", "Atlantic_Media", "The_Verge", "Vox_Media", "Engadget", "AOL", "The_Indian_Express", "Firstpost", "Network_18", "Doi_(identifier)", "ISSN_(identifier)", "S2CID_(identifier)", "Mainframe_server", "VentureBeat", "The_New_York_Times", "The_Wall_Street_Journal", "The_Verge", "Vox_Media", "TechCrunch", "AOL", "ISSN_(identifier)", "ISSN_(identifier)", "ISSN_(identifier)", "The_New_York_Times", "The_New_York_Times", "YouTube", "Wayback_Machine", "Julia_Robinson_Mathematics_Festival", "PC_Magazine", "Ziff_Davis", "Wired_(website)", "Bloomberg_News", "Bloomberg_L.P.", "The_New_York_Times", "BBC_News", "The_Guardian", "PCMag", "ZDNet", "Business_Insider", "Bloomberg_L.P.", "APNews.com", "Newsweek", "USA_Today", "Free_Law_Project", "BBC_News", "The_New_York_Times", "The_Wall_Street_Journal", "ISSN_(identifier)", "The_Wall_Street_Journal", "ISSN_(identifier)", "The_Wall_Street_Journal", "ISSN_(identifier)", "The_Verge", "Bloomberg_News", "Ars_Technica", "Reuters", "Michael_J._Saylor", "ISBN_(identifier)", "Template_talk:Google_LLC", "Alphabet_Inc.", "History_of_Google", "List_of_Android_apps_by_Google", "List_of_Google_Easter_eggs", "List_of_Google_April_Fools%27_Day_jokes", "List_of_mergers_and_acquisitions_by_Alphabet", "Google_Ads", "Google_AI", "Google_Brain", "Android_(operating_system)", "Booting_process_of_Android_devices", "Android_recovery_mode", "Android_software_development", "Android_version_history", "Google_China", "Google_Chrome", "Google_Cloud_Platform", "Google_Glass", "Google.org", "Google_Crisis_Response", "Google_Public_Alerts", "RechargeIT", "Google_Health", "Google_Maps", "Google_Nest", "Google_Pixel", "Google_Search", "Timeline_of_Google_Search", "Sidewalk_Labs", "Sidewalk_Toronto", "Google_Stadia", "YouTube", "History_of_YouTube", "Me_at_the_zoo", "Social_impact_of_YouTube", "YouTuber", "Category:Google_employees", "Krishna_Bharat", "Vint_Cerf", "Jeff_Dean", "John_Doerr", "Sanjay_Ghemawat", "Al_Gore", "John_L._Hennessy", "Urs_H%C3%B6lzle", "Salar_Kamangar", "Ray_Kurzweil", "Ann_Mather", "Alan_Mulally", "Sundar_Pichai", "Ruth_Porat", "Rajen_Sheth", "Hal_Varian", "Susan_Wojcicki", "Andy_Bechtolsheim", "Sergey_Brin", "David_Cheriton", "Matt_Cutts", "David_Drummond_(businessman)", "Alan_Eustace", "Timnit_Gebru", "Omid_Kordestani", "Paul_Otellini", "Larry_Page", "Patrick_Pichette", "Eric_Schmidt", "Ram_Shriram", "Amit_Singhal", "Shirley_M._Tilghman", "Rachel_Whetstone", "111_Eighth_Avenue", "Androidland", "Google_barges", "Chelsea_Market", "Chrome_Zone", "Google_data_centers", "Google_Modular_Data_Center", "Googleplex", "St._John%27s_Terminal", "YouTube_Space", "YouTube_Theater", "Croscore_fonts", "Product_Sans#Variations", "Noto_fonts", "Product_Sans", "Roboto", "Google_logo", "Google_Doodle", "Doodle_Champion_Island_Games", "Magic_Cat_Academy", "Google_logo#Favicon", "Material_Design", "Comparison_of_Material_Design_implementations", "Android_Developer_Challenge", "Android_Developer_Day", "Android_Developer_Lab", "Google_Code-in", "Google_Code_Jam", "Google_Developer_Day", "Google_Developers_Live", "Doodle4Google", "Google_I/O", "Jigsaw_(company)", "Living_Stories", "Google_Lunar_X_Prize", "Google_Mapathon", "Google_Science_Fair", "Google_Summer_of_Code", "Talks_at_Google", "YouTube_Awards", "CNN/YouTube_presidential_debates", "YouTube_Comedy_Week", "YouTube_Live", "YouTube_Music_Awards", "YouTube_Space_Lab", "YouTube_Symphony_Orchestra", "20%25_Project", "A_Google_A_Day", "Area_120", "Reply_(Google)", "Tables_(Google)", "Google_ATAP", "Google_Data_Liberation_Front", "Data_Transfer_Project", "Digital_News_Initiative", "Digital_Unlocked", "Dragonfly_(search_engine)", "Google_Founders%27_Award", "Google_Free_Zone", "Google_Get_Your_Business_Online", "Google_for_Education", "Google_for_Startups", "Google_Labs", "Liquid_Galaxy", "Made_with_Code", "Google_Maori", "Google_News_Lab", "Google_PowerMeter", "Project_Nightingale", "Privacy_Sandbox", "RechargeIT", "Project_Shield", "Project_Starline", "Google_Student_Ambassador_Program", "Dunant_(submarine_communications_cable)", "Grace_Hopper_(submarine_communications_cable)", "Project_Sunroof", "Where_on_Google_Earth_is_Carmen_Sandiego%3F", "YouTube_Creator_Awards", "YouTube_Next_Lab_and_Audience_Development_Group", "YouTube_Original_Channel_Initiative", "Project_Zero", "Criticism_of_Google", "2018_Google_data_breach", "2018_Google_walkouts", "2020_Google_services_outages", "Alphabet_Workers_Union", "Censorship_by_Google", "DeGoogle", "Dragonfly_(search_engine)", "FairSearch", "Google%27s_Ideological_Echo_Chamber", "Is_Google_Making_Us_Stupid%3F", "Google_litigation", "Predictions_of_the_end_of_Google", "Privacy_concerns_regarding_Google", "Google_Street_View_privacy_concerns", "San_Francisco_tech_bus_protests", "Google_worker_organization", "YouTuber_back_advertisement_controversy_in_Korea_in_2020", "Censorship_by_Google#YouTube", "YouTube_copyright_issues", "YouTube_copyright_strike", "Elsagate", "Fantastic_Adventures_scandal", "YouTube_headquarters_shooting", "2012_Kohistan_video_case", "Reactions_to_Innocence_of_Muslims", "2011_Slovenian_YouTube_incident", "Google_Developers", "Android_(operating_system)", "Android_Automotive", "Glass_OS", "Android_Go", "GLinux", "Goobuntu", "Android_Things", "Android_TV", "Wear_OS", "Chrome_OS", "Chromium_OS", "Neverware", "Fuchsia_(operating_system)", "Google_TV_(operating_system)", "Accelerated_Mobile_Pages", "Angular_(web_framework)", "AngularJS", "ARCore", "Google_APIs", "Blockly", "Google_Chart_API", "Google_Charts", "Dialogflow", "Exposure_Notification", "Fast_Pair", "Federated_Learning_of_Cohorts", "Google_File_System", "FlatBuffers", "Flutter_(software)", "Gears_(software)", "GRPC", "Gson", "Google_Guava", "Google_Guice", "Guetzli", "GVisor", "MapReduce", "Google_mobile_services", "Google_Neural_Machine_Translation", "OpenSocial", "Google_Pack", "Polymer_(library)", "Protocol_Buffers", "Reqwireless", "Google_Shell", "Skia_Graphics_Engine", "Tango_(platform)", "TensorFlow", "Google_Test", "WaveNet", "Weave_(protocol)", "Google_Web_Accelerator", "WebRTC", "Google_App_Engine", "Google_Apps_Script", "Google_Cloud_Platform", "Anvato", "Bigtable", "Chronicle_Security", "VirusTotal", "Google_Compute_Engine", "Google_Cloud_Connect", "Google_Cloud_Dataflow", "Google_Cloud_Datastore", "Kaggle", "Looker_(company)", "Google_Cloud_Messaging", "Google_Cloud_Shell", "Stackdriver", "Google_Cloud_Storage", "Firebase", "Firebase_Cloud_Messaging", "Crashlytics", "Global_IP_Solutions", "Internet_Low_Bitrate_Codec", "Internet_Speech_Audio_Codec", "Gridcentric,_Inc.", "ITA_Software", "Kubernetes", "LevelDB", "Orbitera", "SageTV", "Android_Cloud_to_Device_Messaging", "Android_Studio", "App_Inventor_for_Android", "Google_App_Maker", "Google_App_Runtime_for_Chrome", "AppSheet", "Bazel_(software)", "Google_Chrome_Frame", "Google_Closure_Tools", "Cpplint", "Google_Gadgets", "GData", "Gerrit_(software)", "GYP_(software)", "Google_Kythe", "Google_Lighthouse", "Google_Mashup_Editor", "Google_Native_Client", "Google_Optimize", "OpenRefine", "OR-Tools", "Google_PageSpeed_Tools", "Google_Plugin_for_Eclipse", "Google_Programmable_Search_Engine", "Google_Public_DNS", "ReCAPTCHA", "Google_Search_Console", "Sitemaps", "Google_Swiffy", "Tesseract_(software)", "Trendalyzer", "Google_Wave_Federation_Protocol", "Google_Web_Toolkit", "Google_Hummingbird", "PageRank", "Google_matrix", "Google_Panda", "Google_Penguin", "Google_Pigeon", "RankBrain", "BigQuery", "Google_Chrome_Experiments", "Android_App_Bundle", "Android_application_package", "AV1", "On2_Technologies", "VP3", "VP6", "VP8", "Libvpx", "VP9", "WebM", "WebP", "Web_Open_Font_Format", "Open_Location_Code", "Flutter_(American_company)", "Caja_project", "Dart_(programming_language)", "Go_(programming_language)", "Sawzall_(programming_language)", "Viewdle", "Google_Web_Server", "List_of_Google_products", "Google_Currents_(2011%E2%80%932013)", "Green_Throttle_Games", "Owlchemy_Labs", "Oyster_(company)", "PaperofRecord.com", "Google_Play", "Google_Play_Books", "Google_Play_Games", "List_of_most-downloaded_Google_Play_applications", "Google_Play_Music", "Google_Play_Newsstand", "Google_Play_Pass", "Google_Play_Services", "Google_Podcasts", "Quick,_Draw!", "Google_Santa_Tracker", "Google_Stadia", "List_of_Stadia_games", "Typhoon_Studios", "Google_TV_(service)", "Vevo", "Google_Video", "YouTube", "YouTube_API", "BandPage", "Content_ID_(system)", "FameBit", "YouTube_Instant", "YouTube_Kids", "YouTube_Music", "YouTube_(channel)", "Google_Preferred", "YouTube_Premium", "List_of_YouTube_Premium_original_programming", "YouTube_Rewind", "RightsFlow", "YouTube_TV", "Google_Allo", "Bump_(application)", "Google_Buzz", "Google_Chat", "Google_Contacts", "Google_Currents", "Dodgeball_(service)", "Google_Duo", "Google_Fi", "Google_Friend_Connect", "Gizmo5", "Google%2B", "Gmail", "History_of_Gmail", "Inbox_by_Gmail", "Gmail_interface", "Google_Groups", "Google_Hangouts", "Google_Helpouts", "Google_IME", "Google_Japanese_Input", "Google_Pinyin", "Google_transliteration", "Jaiku", "Meebo", "Google_Meet", "Messages_(Google)", "Google_Moderator", "Neotonic_Software", "Orkut", "Postini", "Google_Schemer", "Google_Spaces", "Sparrow_(email_client)", "Google_Talk", "Google_Translate", "Google_Translator_Toolkit", "Google_Voice", "GOOG-411", "Google_Wave", "Word_Lens", "Quest_Visual", "Google_Search", "Aardvark_(search_engine)", "Akwan_Information_Technologies", "Google_Alerts", "Google_Answers", "Google_Base", "Google_Blog_Search", "Google_Books", "Google_Ngram_Viewer", "Google_Code_Search", "Google_Dataset_Search", "Google_Dictionary", "Google_Directory", "Google_Fast_Flip", "Google_Flu_Trends", "Google_Finance", "Google_Goggles", "Googlebot", "Google_Images", "Google_Image_Labeler", "Google_Image_Swirl", "Kaltix", "Google_Knowledge_Graph", "Freebase_(database)", "Metaweb", "Like.com", "Google_News", "Google_News_Archive", "Google_News_%26_Weather", "Google_Patents", "People_Cards", "Google_Personalized_Search", "Google_Public_Data_Explorer", "Google_Questions_and_Answers", "SafeSearch", "Google_Scholar", "Google_SearchWiki", "Google_Shopping", "Google_Catalogs", "Google_Express", "Google_Squared", "Tenor_(website)", "Google_Travel", "Google_Flights", "Google_Trends", "Google_Insights_for_Search", "Google_Voice_Search", "WDYL_(search_engine)", "Google_Earth", "ImageAmerica", "Google_Maps", "Google_Latitude", "Google_Map_Maker", "Google_Maps_Navigation", "Google_Maps_pin", "Google_Street_View", "Coverage_of_Google_Street_View", "Street_View_Trusted", "Waze", "Google_Ad_Manager", "AdMob", "Google_Ads", "Adscape", "Google_AdSense", "Google_Attribution", "BebaPay", "Google_Business_Groups", "Google_Checkout", "Google_Contributor", "DMarc_Broadcasting", "DoubleClick", "Google_Affiliate_Network", "Invite_Media", "Google_Analytics", "Google_Analytics", "Google_Data_Studio", "Urchin_(software)", "Google_Pay", "Google_Pay_Send", "Tez_(software)", "PostRank", "Google_Primer", "Softcard", "Wildfire_Interactive", "Google_Bookmarks", "Google_Browser_Sync", "Google_Calendar", "Google_Cloud_Search", "Google_Desktop", "Google_Docs_Editors", "Google_Docs", "Google_Drawings", "Google_Forms", "Google_Fusion_Tables", "Google_Keep", "Google_Sheets", "Google_Slides", "Google_Sites", "Google_Drive", "Fflick", "Files_(Google)", "IGoogle", "Jamboard", "Google_Notebook", "Google_One", "Google_Photos", "Apture", "Blogger_(service)", "Pyra_Labs", "Google_Domains", "FeedBurner", "Google_One_Pass", "Google_Page_Creator", "Google_Sites", "Google_Web_Designer", "Quickoffice", "Google_Quick_Search_Box", "Google_Surveys", "Google_Sync", "Google_Calendar", "Google_Toolbar", "Google_Account", "Google_Dashboard", "Google_Takeout", "Android_Auto", "Android_Beam", "Google_Arts_%26_Culture", "Google_Assistant", "Google_Authenticator", "ZygoteBody", "BufferBox", "Google_Building_Maker", "Google_Cast", "List_of_apps_with_Google_Cast_support", "Google_Chrome", "Google_Chrome_App", "Chromium_(web_browser)", "Dinosaur_Game", "GreenBorder", "Chrome_Remote_Desktop", "Google_Chrome_version_history", "Chrome_Web_Store", "Google_Classroom", "Google_Cloud_Print", "Crowdsource_(app)", "Google_Expeditions", "Google_Family_Link", "Find_My_Device", "Google_Fit", "Google_Fonts", "Gboard", "Google_Gesture_Search", "Google_Camera", "Google_Lens", "Snapseed", "Nik_Software", "Panoramio", "Google_Photos", "Picasa", "Picasa_Web_Albums", "Picnik", "Impermium", "Knol", "Google_Lively", "Live_Transcribe", "MyTracks", "Nearby_Share", "Google_Now", "Google_Offers", "Google_Opinion_Rewards", "Google_Person_Finder", "PlinkArt", "Poly_(website)", "Google_Question_Hub", "Read_Along", "Google_Reader", "Google_Safe_Browsing", "Google_Sidewiki", "SlickLogin", "Socratic_(Google)", "Sound_Amplifier", "Speech_Services", "Google_Station", "Google_Store", "Google_TalkBack", "Tilt_Brush", "Google_URL_Shortener", "Wavii", "Google_Web_Light", "Google_WiFi", "Google_Workspace", "Google_Workspace_Marketplace", "Android_One", "Google_Nexus", "Nexus_One", "Nexus_S", "Galaxy_Nexus", "Nexus_4", "Nexus_5", "Nexus_6", "Nexus_5X", "Nexus_6P", "Comparison_of_Google_Nexus_smartphones", "Google_Pixel", "Pixel_(1st_generation)", "Pixel_2", "Pixel_3", "Pixel_3a", "Pixel_4", "Pixel_4a", "Pixel_5", "Pixel_5a", "Pixel_6", "Comparison_of_Google_Pixel_smartphones", "List_of_Google_Play_edition_devices", "Project_Ara", "Chromebook", "Google_Nexus", "Nexus_7_(2012)", "Nexus_7_(2013)", "Nexus_10", "Nexus_9", "Comparison_of_Google_Nexus_tablets", "Google_Pixel", "Chromebook_Pixel", "Pixelbook", "Pixelbook_Go", "Pixel_C", "Pixel_Slate", "Chromebit", "Chromebox", "Google_Clips", "Chromecast", "Nexus_Player", "Nexus_Q", "Dropcam", "Fitbit", "List_of_Fitbit_products", "Google_Nest", "Google_Nest#Nest_Cam_Indoor", "Google_Nest#Nest_Doorbell", "Google_Nest#Nest_Protect", "Google_Nest#Nest_Secure", "Google_Nest_(smart_speakers)", "Nest_Thermostat", "Google_Nest_Wifi", "Google_OnHub", "Pixel_Buds", "Google_Pixel#Pixel_Stand", "Pixel_Visual_Core", "Google_Search_Appliance", "Sycamore_processor", "Tensor_Processing_Unit", "Google_Cardboard", "Google_Contact_Lens", "Google_Daydream", "Google_Glass", "Don%27t_be_evil", "Gayglers", "Google_(verb)", "Googlearchy", "Google_bombing", "Google_effect", "Googlefight", "Google_hacking", "Googleshare", "Google_tax", "Googlewhack", "Googlization", "Rooting_(Android)", "YouTube_Poop", "Category:Documentary_films_about_Google", "Google_Behind_the_Screen", "Google_Current", "Google_Maps_Road_Trip", "Google_The_Thinking_Factory", "Google_and_the_World_Brain", "The_Creepy_Line", "Category:Books_about_Google", "Google_Hacks", "The_Google_Story", "Google_Volume_One", "Googled:_The_End_of_the_World_as_We_Know_It", "How_Google_Works", "I%27m_Feeling_Lucky_(book)", "In_the_Plex", ".google", "Google_Developer_Expert", "Google_Feud", "Google_Me_(film)", "Google_Me_(song)", "The_Billion_Dollar_Code", "The_Internship", "List_of_Google_products#Discontinued_products_and_services", "Category:Google", "Outline_of_Google", "Template_talk:Alphabet_Inc.", "Alphabet_Inc.", "Google_Fiber", "Calico_(company)", "CapitalG", "DeepMind", "AlphaFold", "AlphaGo", "AlphaGo_Zero", "AlphaStar_(software)", "AlphaZero", "Master_(software)", "MuZero", "WaveNet", "DoubleClick", "Firebase", "Fitbit", "Tenor_(website)", "Waze", "YouTube", "GV_(company)", "Isomorphic_Labs", "Verily", "X_Development", "Waymo", "Wing_(company)", "Boston_Dynamics", "Chronicle_Security", "Jigsaw_(company)", "Loon_LLC", "Makani_(company)", "Meka_Robotics", "Google_Nest", "Redwood_Robotics", "Sidewalk_Labs", "Sidewalk_Toronto", "Sundar_Pichai", "Chief_executive_officer", "Ruth_Porat", "Chief_financial_officer", "Larry_Page", "Chief_executive_officer", "Sergey_Brin", "President_(corporate_title)", "David_Drummond_(businessman)", "Chief_legal_officer", "Sergey_Brin", "John_Doerr", "John_L._Hennessy", "Ann_Mather", "Alan_Mulally", "Larry_Page", "Sundar_Pichai", "Ram_Shriram", "Roger_W._Ferguson_Jr.", "Eric_Schmidt", "Diane_Greene", "Andrew_Conrad", "Tony_Fadell", "Arthur_D._Levinson", "David_Krane", "Ruth_Porat", "Astro_Teller", "Category:Alphabet_Inc.", "Portal:Companies", "Portal:Internet", "Template_talk:Open_Handset_Alliance_Members", "Open_Handset_Alliance", "Mobile_network_operator", "Bouygues_Telecom", "China_Mobile", "China_Telecommunications_Corporation", "China_Unicom", "Gruppo_TIM", "KDDI", "Nepal_Telecom", "NTT_Docomo", "SoftBank_Group", "Sprint_Corporation", "T-Mobile", "Telef%C3%B3nica", "Telus", "Vodafone", "Software", "Access_(company)", "Ascender_Corporation", "EBay", "Myriad_Group", "Nuance_Communications", "NXP_Semiconductors", "Omron", "PacketVideo", "SVOX", "VisualOn", "Semiconductor", "AKM_Semiconductor", "Arm_Holdings", "Audience_(company)", "Broadcom_Corporation", "CSR_plc", "SiRF", "Cypress_Semiconductor", "Freescale_Semiconductor", "Gemalto", "Intel", "Marvell_Technology_Group", "MediaTek", "MIPS_Technologies", "Nvidia", "Qualcomm", "Qualcomm_Atheros", "Renesas_Electronics", "ST-Ericsson", "Ericsson_Mobile_Platforms", "Synaptics", "Texas_Instruments", "Mobile_phone", "Acer_Inc", "Alcatel_Mobile_Phones", "Asus", "Chaudhary_Group", "Compal_Electronics", "Dell", "Foxconn", "Garmin", "HTC", "Huawei", "Kyocera", "Lenovo", "LG_Electronics", "Motorola_Mobility", "NEC", "Samsung_Electronics", "Sharp_Corporation", "Sony_Mobile", "Toshiba", "ZTE", "Marketing", "Accenture", "Borqs", "Sasken_Communication_Technologies", "Teleca", "The_Astonishing_Tribe", "Wind_River_Systems", "Wipro_Technologies", "Android_(operating_system)", "Dalvik_(software)", "Google_Nexus", "HTC_Dream", "Template_talk:Google_Lunar_X_Prize", "Google_Lunar_X_Prize", "Sundar_Pichai", "Larry_Page", "Sergey_Brin", "X_Prize_Foundation", "Peter_Diamandis", "Hakuto", "Moon_Express", "SpaceIL", "Synergy_Moon", "TeamIndus", "Team_AngelicvM", "ARCA_Space_Corporation", "Astrobotic_Technology", "Barcelona_Moon_Team", "Euroluna", "Team_FREDNET", "Independence-X_Aerospace", "Micro-Space", "Odyssey_Moon", "PTScientists", "Penn_State_Lunar_Lion_Team", "Puli_Space_Technologies", "Rocket_City_Space_Pioneers", "Selenokhod", "SpaceMETA", "PTScientists", "Beresheet", "TeamIndus", "MX-1E", "Astrobotic", "Hakuto", "Synergy_Moon", "Team_AngelicvM", "Template_talk:Major_mobile_device_companies", "Mobile_device", "Acer_Inc.", "Alba_(brand)", "Amazon_(company)", "Apple_Inc.", "Asus", "BBK_Electronics", "Oppo", "OnePlus", "Vivo_(technology_company)", "Realme", "IQOO", "BlackBerry_Limited", "BLU_Products", "Bush_(brand)", "Dell", "Fitbit", "Foxconn", "Sharp_Corporation", "InFocus", "Nokia", "Fujitsu", "GeeksPhone", "Gionee", "Haier", "Hisense", "Honor_(brand)", "HTC", "HP_Inc.", "Huawei", "Intel", "Kyocera", "Lenovo", "Motorola_Mobility", "LG_Electronics", "Meizu", "Microsoft", "Microsoft_Lumia", "HMD_Global", "Nokia", "Onyx_Boox", "Panasonic", "Samsung_Electronics", "Sony", "TCL_Technology", "Alcatel_Mobile", "BlackBerry_Mobile", "RCA", "Palm,_Inc.", "Toshiba", "Transsion", "Tecno_Mobile", "Infinix_Mobile", "Itel_Mobile", "Tinno_Mobile", "Wiko", "True_Corporation", "Vaio", "VinSmart", "Xiaomi", "Redmi", "POCO_(company)", "ZTE", "Nubia_Technology", "Walton_Micro-Tech_Corporation", "List_of_largest_technology_companies_by_revenue", "Category:Mobile_technology_companies", "Category:Mobile_phone_manufacturers", "Template_talk:Major_Internet_companies", "Internet", "List_of_largest_Internet_companies", "Adobe_Inc.", "Alibaba_Group", "Alphabet_Inc.", "Amazon_(company)", "Apple_Inc.", "Baidu", "IAC_(company)", "Meituan", "Meta_Platforms", "Facebook", "Microsoft", "Naver_Corporation", "NetEase", "Tencent", "Yandex", "Akamai_Technologies", "Alibaba_Cloud", "Amazon_Web_Services", "ICloud", "Google_Cloud_Platform", "IBM_cloud_computing", "Microsoft_Azure", "Oracle_Corporation", "Rackspace_Technology", "Salesforce", "ServiceNow", "Amazon_(company)", "Apple_Inc.", "Booking_Holdings", "Coupang", "EBay", "Expedia", "Flipkart", "Groupon", "JD.com", "Lazada_Group", "Rakuten", "Shopee", "Shopify", "Suning.com", "Trip.com", "Uber", "Wayfair", "Zalando", "Bloomberg_L.P.", "BuzzFeed", "ByteDance", "Kuaishou", "G/O_Media", "Netflix", "Spotify", "ViacomCBS_Streaming", "Vox_Media", "Yahoo_(company)", "List_of_largest_technology_companies_by_revenue", "List_of_largest_Internet_companies", "Category:Internet_technology_companies", "Template_talk:Major_software_companies", "Software", "Adobe_Inc", "Amadeus_IT_Group", "Amazon_(company)", "Apple_Inc", "Autodesk", "Citrix_Systems", "FIS_(company)", "Hewlett_Packard_Enterprise", "IBM", "Intuit", "Infor", "Microsoft", "Oracle_Corporation", "Quest_Software", "Sage_Group", "SAP", "Tencent", "List_of_largest_technology_companies_by_revenue", "List_of_the_largest_software_companies", "Category:Software_companies", "Template_talk:Princess_of_Asturias_Award_for_Communication_and_Humanities", "Princess_of_Asturias_Awards", "Mar%C3%ADa_Zambrano", "Mario_Bunge", "El_Pa%C3%ADs", "Claudio_S%C3%A1nchez-Albornoz", "Jos%C3%A9_Ferrater_Mora", "Grupo_Globo", "El_Espectador", "El_Tiempo_(Colombia)", "Pedro_La%C3%ADn_Entralgo", "Fondo_de_Cultura_Econ%C3%B3mica", "Central_American_University", "Emilio_Garc%C3%ADa_G%C3%B3mez", "Vuelta_(magazine)", "Octavio_Paz", "Mission_(Christian)", "Rwanda", "Burundi", "EFE", "Indro_Montanelli", "Juli%C3%A1n_Mar%C3%ADas", "V%C3%A1clav_Havel", "CNN", "Reinhard_Mohn", "Caro_and_Cuervo_Institute", "Umberto_Eco", "George_Steiner", "Hans_Magnus_Enzensberger", "Ryszard_Kapu%C5%9Bci%C5%84ski", "Gustavo_Guti%C3%A9rrez", "Jean_Daniel", "Alliance_Fran%C3%A7aise", "Dante_Alighieri_Society", "British_Council", "Goethe-Institut", "Instituto_Cervantes", "Instituto_Cam%C3%B5es", "National_Geographic_Society", "Nature_(journal)", "Science_(journal)", "National_Autonomous_University_of_Mexico", "Alain_Touraine", "Zygmunt_Bauman", "Royal_Society", "Shigeru_Miyamoto", "Annie_Leibovitz", "Quino", "Emilio_Lled%C3%B3", "James_Nachtwey", "Les_Luthiers", "Alma_Guillermoprieto", "Museo_del_Prado", "Guadalajara_International_Book_Fair", "Hay_Festival", "Gloria_Steinem", "Template_talk:Electronics_industry_in_the_United_States", "Electronics_industry_in_the_United_States", "Home_appliance", "Apple_Inc.", "Bose_Corporation", "Cisco_Systems", "Corsair_Gaming", "Dell", "Dolby_Laboratories", "Element_Electronics", "Emerson_Radio", "Harman_International_Industries", "JBL_(company)", "Honeywell", "HP_Inc.", "InFocus", "Jensen_Electronics", "Kenmore_(brand)", "Kingston_Technology", "Kimball_International", "Koss_Corporation", "Lexmark", "Logitech", "Magnavox", "Marantz", "Memorex", "Microsoft", "Monster_Cable", "Plantronics", "Planar_Systems", "Pyle_USA", "Razer_Inc.", "Seagate_Technology", "Seiki_Digital", "Skullcandy", "Turtle_Beach_Corporation", "ViewSonic", "Vizio_Inc.", "Western_Digital", "HGST", "SanDisk", "Westinghouse_Electric_Company", "Westinghouse_Electronics", "Xerox", "Electronic_component", "3M", "Achronix", "Analog_Devices", "Maxim_Integrated", "Applied_Materials", "Altera", "AVX_Corporation", "Cirque_Corporation", "Diodes_Incorporated", "Flex_(company)", "Jabil", "KEMET_Corporation", "Maxwell_Technologies", "Sanmina_Corporation", "Vishay_Intertechnology", "Semiconductor_device", "Advanced_Micro_Devices", "Ampere_Computing", "Apple_silicon", "Broadcom_Inc.", "Cypress_Semiconductor", "GlobalFoundries", "IBM", "Intel", "Interlink_Electronics", "KLA-Tencor", "Lam_Research", "Lattice_Semiconductor", "Marvell_Technology,_Inc.", "Microchip_Technology", "Atmel", "Micron_Technology", "NetApp", "Nimbus_Data", "Nvidia", "Mellanox_Technologies", "NXP_Semiconductors", "ON_Semiconductor", "Qualcomm", "Silicon_Image", "Synaptics", "Tabula_(company)", "Texas_Instruments", "Xilinx", "Zilog", "Mobile_device", "Apple_Inc.", "BLU_Products", "Lenovo", "Motorola_Mobility", "Cadence_Design_Systems", "Cray", "General_Electric", "RCA_(trademark)", "Oracle_Corporation", "Synopsys", "Actel", "Atari_Corporation", "Commodore_International", "Compaq", "Fairchild_Semiconductor", "Freescale_Semiconductor", "LSI_Corporation", "Microsemi", "National_Semiconductor", "Palm,_Inc.", "Philco", "RCA", "Signetics", "Silicon_Graphics", "Solectron", "Sun_Microsystems", "Zenith_Electronics", "Portal:Companies", "Portal:Computer_programming", "Portal:Internet", "Portal:San_Francisco_Bay_Area", "Portal:Technology", "ISNI_(identifier)", "VIAF_(identifier)", "SUDOC_(identifier)"], "Alphabet_Inc.": ["Google", "Googleplex", "Mountain_View,_California", "Public_company", "Ticker_symbol", "Nasdaq", "Nasdaq-100", "S%26P_100", "S%26P_500", "International_Securities_Identification_Number", "Conglomerate_(company)", "Larry_Page", "Sergey_Brin", "Googleplex", "Mountain_View,_California", "U.S.", "John_L._Hennessy", "Sundar_Pichai", "Ruth_Porat", "Artificial_intelligence", "Automation", "Autonomous_car", "Biotechnology", "Cloud_computing", "Computer_hardware", "Corporate_venture_capital", "Fiber_to_the_x", "Health_care", "Internet", "Robotics", "Software", "United_States_dollar", "Earnings_before_interest_and_taxes", "Net_income", "Asset", "Equity_(finance)", "Larry_Page", "Sergey_Brin", "Subsidiary", "Calico_(company)", "CapitalG", "DeepMind", "Google", "Google_Fiber", "GV_(company)", "Intrinsic_(company)", "Isomorphic_Labs", "Verily", "Waymo", "Wing_(company)", "X_Development", "Conglomerate_(company)", "Holding_company", "Mountain_View,_California", "Restructuring", "Google", "Subsidiary", "List_of_largest_technology_companies_by_revenue", "List_of_public_corporations_by_market_capitalization", "Larry_Page", "Sergey_Brin", "Sundar_Pichai", "Controlling_interest", "Google", "Holding_company", "Larry_Page", "X_(company)", "Calico_(company)", "Google_Nest", "Verily", "Google_Fiber", "Makani_(company)", "CapitalG", "GV_(company)", "Sundar_Pichai", "Larry_Page", "Warren_Buffett", "Berkshire_Hathaway", "Delaware_General_Corporation_Law", "Stock_market_index", "S%26P_500", "NASDAQ-100", "Sundar_Pichai", "Calico_(company)", "Arthur_D._Levinson", "CapitalG", "Private_equity", "David_Lawee", "DeepMind", "Artificial_intelligence", "Demis_Hassabis", "Google", "Sundar_Pichai", "Google_Fiber", "GV_(company)", "Venture_capital", "David_Krane", "Intrinsic_(company)", "Robotics", "Wendy_Tan_White", "Isomorphic_Labs", "Drug_discovery", "Verily_Life_Sciences", "Andrew_Conrad", "Waymo", "Dmitri_Dolgov", "Tekedra_Mawakana", "Wing_(company)", "X_Development", "Astro_Teller", "Limited_liability_company", "Eric_Schmidt", "Internet_Association", "Android_(operating_system)", "YouTube", "Google_Search", "Nest_Labs", "Chronicle_(company)", "Sidewalk_Labs", "Daniel_L._Doctoroff", "ALS", "Loon_LLC", "Intrinsic_(company)", "Robotics", "Isomorphic_Labs", "Drug_discovery", "Demis_Hassabis", "Alpha_(finance)", ".xyz", "Top-level_domain", "Fleet_management", "BMW", "The_Walt_Disney_Company", "American_Broadcasting_Company", "Easter_egg_(media)", "Google_X_Labs", "X_(company)#Wing", "Drone_delivery", "Hyperlink", "Silicon_Valley_(TV_series)", "AdSense", "Google_Ads", "Apple_Inc.", "Fortune_500", "Trillion", "List_of_public_corporations_by_market_capitalization", "Andreessen_Horowitz", "20th_Century_Fox", "UnitedMasters", "Steve_Stoute", "Uber", "Medium_(website)", "List_of_mergers_and_acquisitions_by_Alphabet", "Intel", "Cisco", "Hoffmann-La_Roche", "Oncology", "Uber", "Self-driving_car", "Autonomous_vehicle", "Waymo", "Proprietary_software", "Waymo", "Google", "United_States_Department_of_Justice", "National_Labor_Relations_Board", "Antitrust", "Autorit%C3%A9_de_la_concurrence", "Google_LLC", "Portal:San_Francisco_Bay_Area", "Portal:Internet", "U.S._Securities_and_Exchange_Commission", "Wired_(website)", "Wired_(website)", "United_States_Securities_and_Exchange_Commission", "The_Verge", "CNBC", "CNBC", "The_Daily_Telegraph", "TechCrunch", "David_Stubbs", "The_Guardian", "CNBC", "NBCUniversal", "USA_Today", "Fortune_(magazine)", "Techcrunch", "The_New_York_Times", "CNBC", "Newsweek", "USA_Today", "ISSN_(identifier)", "Template_talk:Alphabet_Inc.", "Google_Fiber", "Calico_(company)", "CapitalG", "DeepMind", "AlphaFold", "AlphaGo", "AlphaGo_Zero", "AlphaStar_(software)", "AlphaZero", "Master_(software)", "MuZero", "WaveNet", "Google", "DoubleClick", "Firebase", "Fitbit", "Tenor_(website)", "Waze", "YouTube", "GV_(company)", "Isomorphic_Labs", "Verily", "X_Development", "Waymo", "Wing_(company)", "Boston_Dynamics", "Chronicle_Security", "Jigsaw_(company)", "Loon_LLC", "Makani_(company)", "Meka_Robotics", "Google_Nest", "Redwood_Robotics", "Sidewalk_Labs", "Sidewalk_Toronto", "Sundar_Pichai", "Chief_executive_officer", "Ruth_Porat", "Chief_financial_officer", "Larry_Page", "Chief_executive_officer", "Sergey_Brin", "President_(corporate_title)", "David_Drummond_(businessman)", "Chief_legal_officer", "Sergey_Brin", "John_Doerr", "John_L._Hennessy", "Ann_Mather", "Alan_Mulally", "Larry_Page", "Sundar_Pichai", "Ram_Shriram", "Roger_W._Ferguson_Jr.", "Eric_Schmidt", "Diane_Greene", "Andrew_Conrad", "Tony_Fadell", "Arthur_D._Levinson", "David_Krane", "Ruth_Porat", "Astro_Teller", "Category:Alphabet_Inc.", "Portal:Companies", "Portal:Internet", "Template_talk:Google_LLC", "Google", "History_of_Google", "List_of_Android_apps_by_Google", "List_of_Google_Easter_eggs", "List_of_Google_April_Fools%27_Day_jokes", "List_of_mergers_and_acquisitions_by_Alphabet", "Google_Ads", "Google_AI", "Google_Brain", "Android_(operating_system)", "Booting_process_of_Android_devices", "Android_recovery_mode", "Android_software_development", "Android_version_history", "Google_China", "Google_Chrome", "Google_Cloud_Platform", "Google_Glass", "Google.org", "Google_Crisis_Response", "Google_Public_Alerts", "RechargeIT", "Google_Health", "Google_Maps", "Google_Nest", "Google_Pixel", "Google_Search", "Timeline_of_Google_Search", "Sidewalk_Labs", "Sidewalk_Toronto", "Google_Stadia", "YouTube", "History_of_YouTube", "Me_at_the_zoo", "Social_impact_of_YouTube", "YouTuber", "Category:Google_employees", "Krishna_Bharat", "Vint_Cerf", "Jeff_Dean", "John_Doerr", "Sanjay_Ghemawat", "Al_Gore", "John_L._Hennessy", "Urs_H%C3%B6lzle", "Salar_Kamangar", "Ray_Kurzweil", "Ann_Mather", "Alan_Mulally", "Sundar_Pichai", "Ruth_Porat", "Rajen_Sheth", "Hal_Varian", "Susan_Wojcicki", "Andy_Bechtolsheim", "Sergey_Brin", "David_Cheriton", "Matt_Cutts", "David_Drummond_(businessman)", "Alan_Eustace", "Timnit_Gebru", "Omid_Kordestani", "Paul_Otellini", "Larry_Page", "Patrick_Pichette", "Eric_Schmidt", "Ram_Shriram", "Amit_Singhal", "Shirley_M._Tilghman", "Rachel_Whetstone", "111_Eighth_Avenue", "Androidland", "Google_barges", "Chelsea_Market", "Chrome_Zone", "Google_data_centers", "Google_Modular_Data_Center", "Googleplex", "St._John%27s_Terminal", "YouTube_Space", "YouTube_Theater", "Croscore_fonts", "Product_Sans#Variations", "Noto_fonts", "Product_Sans", "Roboto", "Google_logo", "Google_Doodle", "Doodle_Champion_Island_Games", "Magic_Cat_Academy", "Google_logo#Favicon", "Material_Design", "Comparison_of_Material_Design_implementations", "Android_Developer_Challenge", "Android_Developer_Day", "Android_Developer_Lab", "Google_Code-in", "Google_Code_Jam", "Google_Developer_Day", "Google_Developers_Live", "Doodle4Google", "Google_I/O", "Jigsaw_(company)", "Living_Stories", "Google_Lunar_X_Prize", "Google_Mapathon", "Google_Science_Fair", "Google_Summer_of_Code", "Talks_at_Google", "YouTube_Awards", "CNN/YouTube_presidential_debates", "YouTube_Comedy_Week", "YouTube_Live", "YouTube_Music_Awards", "YouTube_Space_Lab", "YouTube_Symphony_Orchestra", "20%25_Project", "A_Google_A_Day", "Area_120", "Reply_(Google)", "Tables_(Google)", "Google_ATAP", "Google_Data_Liberation_Front", "Data_Transfer_Project", "Digital_News_Initiative", "Digital_Unlocked", "Dragonfly_(search_engine)", "Google_Founders%27_Award", "Google_Free_Zone", "Google_Get_Your_Business_Online", "Google_for_Education", "Google_for_Startups", "Google_Labs", "Liquid_Galaxy", "Made_with_Code", "Google_Maori", "Google_News_Lab", "Google_PowerMeter", "Project_Nightingale", "Privacy_Sandbox", "RechargeIT", "Project_Shield", "Project_Starline", "Google_Student_Ambassador_Program", "Dunant_(submarine_communications_cable)", "Grace_Hopper_(submarine_communications_cable)", "Project_Sunroof", "Where_on_Google_Earth_is_Carmen_Sandiego%3F", "YouTube_Creator_Awards", "YouTube_Next_Lab_and_Audience_Development_Group", "YouTube_Original_Channel_Initiative", "Project_Zero", "Criticism_of_Google", "2018_Google_data_breach", "2018_Google_walkouts", "2020_Google_services_outages", "Alphabet_Workers_Union", "Censorship_by_Google", "DeGoogle", "Dragonfly_(search_engine)", "FairSearch", "Google%27s_Ideological_Echo_Chamber", "Is_Google_Making_Us_Stupid%3F", "Google_litigation", "Predictions_of_the_end_of_Google", "Privacy_concerns_regarding_Google", "Google_Street_View_privacy_concerns", "San_Francisco_tech_bus_protests", "Google_worker_organization", "YouTuber_back_advertisement_controversy_in_Korea_in_2020", "Censorship_by_Google#YouTube", "YouTube_copyright_issues", "YouTube_copyright_strike", "Elsagate", "Fantastic_Adventures_scandal", "YouTube_headquarters_shooting", "2012_Kohistan_video_case", "Reactions_to_Innocence_of_Muslims", "2011_Slovenian_YouTube_incident", "Google_Developers", "Android_(operating_system)", "Android_Automotive", "Glass_OS", "Android_Go", "GLinux", "Goobuntu", "Android_Things", "Android_TV", "Wear_OS", "Chrome_OS", "Chromium_OS", "Neverware", "Fuchsia_(operating_system)", "Google_TV_(operating_system)", "Accelerated_Mobile_Pages", "Angular_(web_framework)", "AngularJS", "ARCore", "Google_APIs", "Blockly", "Google_Chart_API", "Google_Charts", "Dialogflow", "Exposure_Notification", "Fast_Pair", "Federated_Learning_of_Cohorts", "Google_File_System", "FlatBuffers", "Flutter_(software)", "Gears_(software)", "GRPC", "Gson", "Google_Guava", "Google_Guice", "Guetzli", "GVisor", "MapReduce", "Google_mobile_services", "Google_Neural_Machine_Translation", "OpenSocial", "Google_Pack", "Polymer_(library)", "Protocol_Buffers", "Reqwireless", "Google_Shell", "Skia_Graphics_Engine", "Tango_(platform)", "TensorFlow", "Google_Test", "WaveNet", "Weave_(protocol)", "Google_Web_Accelerator", "WebRTC", "Google_App_Engine", "Google_Apps_Script", "Google_Cloud_Platform", "Anvato", "Bigtable", "Chronicle_Security", "VirusTotal", "Google_Compute_Engine", "Google_Cloud_Connect", "Google_Cloud_Dataflow", "Google_Cloud_Datastore", "Kaggle", "Looker_(company)", "Google_Cloud_Messaging", "Google_Cloud_Shell", "Stackdriver", "Google_Cloud_Storage", "Firebase", "Firebase_Cloud_Messaging", "Crashlytics", "Global_IP_Solutions", "Internet_Low_Bitrate_Codec", "Internet_Speech_Audio_Codec", "Gridcentric,_Inc.", "ITA_Software", "Kubernetes", "LevelDB", "Orbitera", "SageTV", "Android_Cloud_to_Device_Messaging", "Android_Studio", "App_Inventor_for_Android", "Google_App_Maker", "Google_App_Runtime_for_Chrome", "AppSheet", "Bazel_(software)", "Google_Chrome_Frame", "Google_Closure_Tools", "Cpplint", "Google_Gadgets", "GData", "Gerrit_(software)", "GYP_(software)", "Google_Kythe", "Google_Lighthouse", "Google_Mashup_Editor", "Google_Native_Client", "Google_Optimize", "OpenRefine", "OR-Tools", "Google_PageSpeed_Tools", "Google_Plugin_for_Eclipse", "Google_Programmable_Search_Engine", "Google_Public_DNS", "ReCAPTCHA", "Google_Search_Console", "Sitemaps", "Google_Swiffy", "Tesseract_(software)", "Trendalyzer", "Google_Wave_Federation_Protocol", "Google_Web_Toolkit", "Google_Hummingbird", "PageRank", "Google_matrix", "Google_Panda", "Google_Penguin", "Google_Pigeon", "RankBrain", "BigQuery", "Google_Chrome_Experiments", "Android_App_Bundle", "Android_application_package", "AV1", "On2_Technologies", "VP3", "VP6", "VP8", "Libvpx", "VP9", "WebM", "WebP", "Web_Open_Font_Format", "Open_Location_Code", "Flutter_(American_company)", "Caja_project", "Dart_(programming_language)", "Go_(programming_language)", "Sawzall_(programming_language)", "Viewdle", "Google_Web_Server", "List_of_Google_products", "Google_Currents_(2011%E2%80%932013)", "Green_Throttle_Games", "Owlchemy_Labs", "Oyster_(company)", "PaperofRecord.com", "Google_Play", "Google_Play_Books", "Google_Play_Games", "List_of_most-downloaded_Google_Play_applications", "Google_Play_Music", "Google_Play_Newsstand", "Google_Play_Pass", "Google_Play_Services", "Google_Podcasts", "Quick,_Draw!", "Google_Santa_Tracker", "Google_Stadia", "List_of_Stadia_games", "Typhoon_Studios", "Google_TV_(service)", "Vevo", "Google_Video", "YouTube", "YouTube_API", "BandPage", "Content_ID_(system)", "FameBit", "YouTube_Instant", "YouTube_Kids", "YouTube_Music", "YouTube_(channel)", "Google_Preferred", "YouTube_Premium", "List_of_YouTube_Premium_original_programming", "YouTube_Rewind", "RightsFlow", "YouTube_TV", "Google_Allo", "Bump_(application)", "Google_Buzz", "Google_Chat", "Google_Contacts", "Google_Currents", "Dodgeball_(service)", "Google_Duo", "Google_Fi", "Google_Friend_Connect", "Gizmo5", "Google%2B", "Gmail", "History_of_Gmail", "Inbox_by_Gmail", "Gmail_interface", "Google_Groups", "Google_Hangouts", "Google_Helpouts", "Google_IME", "Google_Japanese_Input", "Google_Pinyin", "Google_transliteration", "Jaiku", "Meebo", "Google_Meet", "Messages_(Google)", "Google_Moderator", "Neotonic_Software", "Orkut", "Postini", "Google_Schemer", "Google_Spaces", "Sparrow_(email_client)", "Google_Talk", "Google_Translate", "Google_Translator_Toolkit", "Google_Voice", "GOOG-411", "Google_Wave", "Word_Lens", "Quest_Visual", "Google_Search", "Aardvark_(search_engine)", "Akwan_Information_Technologies", "Google_Alerts", "Google_Answers", "Google_Base", "Google_Blog_Search", "Google_Books", "Google_Ngram_Viewer", "Google_Code_Search", "Google_Dataset_Search", "Google_Dictionary", "Google_Directory", "Google_Fast_Flip", "Google_Flu_Trends", "Google_Finance", "Google_Goggles", "Googlebot", "Google_Images", "Google_Image_Labeler", "Google_Image_Swirl", "Kaltix", "Google_Knowledge_Graph", "Freebase_(database)", "Metaweb", "Like.com", "Google_News", "Google_News_Archive", "Google_News_%26_Weather", "Google_Patents", "People_Cards", "Google_Personalized_Search", "Google_Public_Data_Explorer", "Google_Questions_and_Answers", "SafeSearch", "Google_Scholar", "Google_SearchWiki", "Google_Shopping", "Google_Catalogs", "Google_Express", "Google_Squared", "Tenor_(website)", "Google_Travel", "Google_Flights", "Google_Trends", "Google_Insights_for_Search", "Google_Voice_Search", "WDYL_(search_engine)", "Google_Earth", "ImageAmerica", "Google_Maps", "Google_Latitude", "Google_Map_Maker", "Google_Maps_Navigation", "Google_Maps_pin", "Google_Street_View", "Coverage_of_Google_Street_View", "Street_View_Trusted", "Waze", "Google_Ad_Manager", "AdMob", "Google_Ads", "Adscape", "Google_AdSense", "Google_Attribution", "BebaPay", "Google_Business_Groups", "Google_Checkout", "Google_Contributor", "DMarc_Broadcasting", "DoubleClick", "Google_Affiliate_Network", "Invite_Media", "Google_Analytics", "Google_Analytics", "Google_Data_Studio", "Urchin_(software)", "Google_Pay", "Google_Pay_Send", "Tez_(software)", "PostRank", "Google_Primer", "Softcard", "Wildfire_Interactive", "Google_Bookmarks", "Google_Browser_Sync", "Google_Calendar", "Google_Cloud_Search", "Google_Desktop", "Google_Docs_Editors", "Google_Docs", "Google_Drawings", "Google_Forms", "Google_Fusion_Tables", "Google_Keep", "Google_Sheets", "Google_Slides", "Google_Sites", "Google_Drive", "Fflick", "Files_(Google)", "IGoogle", "Jamboard", "Google_Notebook", "Google_One", "Google_Photos", "Apture", "Blogger_(service)", "Pyra_Labs", "Google_Domains", "FeedBurner", "Google_One_Pass", "Google_Page_Creator", "Google_Sites", "Google_Web_Designer", "Quickoffice", "Google_Quick_Search_Box", "Google_Surveys", "Google_Sync", "Google_Calendar", "Google_Toolbar", "Google_Account", "Google_Dashboard", "Google_Takeout", "Android_Auto", "Android_Beam", "Google_Arts_%26_Culture", "Google_Assistant", "Google_Authenticator", "ZygoteBody", "BufferBox", "Google_Building_Maker", "Google_Cast", "List_of_apps_with_Google_Cast_support", "Google_Chrome", "Google_Chrome_App", "Chromium_(web_browser)", "Dinosaur_Game", "GreenBorder", "Chrome_Remote_Desktop", "Google_Chrome_version_history", "Chrome_Web_Store", "Google_Classroom", "Google_Cloud_Print", "Crowdsource_(app)", "Google_Expeditions", "Google_Family_Link", "Find_My_Device", "Google_Fit", "Google_Fonts", "Gboard", "Google_Gesture_Search", "Google_Camera", "Google_Lens", "Snapseed", "Nik_Software", "Panoramio", "Google_Photos", "Picasa", "Picasa_Web_Albums", "Picnik", "Impermium", "Knol", "Google_Lively", "Live_Transcribe", "MyTracks", "Nearby_Share", "Google_Now", "Google_Offers", "Google_Opinion_Rewards", "Google_Person_Finder", "PlinkArt", "Poly_(website)", "Google_Question_Hub", "Read_Along", "Google_Reader", "Google_Safe_Browsing", "Google_Sidewiki", "SlickLogin", "Socratic_(Google)", "Sound_Amplifier", "Speech_Services", "Google_Station", "Google_Store", "Google_TalkBack", "Tilt_Brush", "Google_URL_Shortener", "Wavii", "Google_Web_Light", "Google_WiFi", "Google_Workspace", "Google_Workspace_Marketplace", "Android_One", "Google_Nexus", "Nexus_One", "Nexus_S", "Galaxy_Nexus", "Nexus_4", "Nexus_5", "Nexus_6", "Nexus_5X", "Nexus_6P", "Comparison_of_Google_Nexus_smartphones", "Google_Pixel", "Pixel_(1st_generation)", "Pixel_2", "Pixel_3", "Pixel_3a", "Pixel_4", "Pixel_4a", "Pixel_5", "Pixel_5a", "Pixel_6", "Comparison_of_Google_Pixel_smartphones", "List_of_Google_Play_edition_devices", "Project_Ara", "Chromebook", "Google_Nexus", "Nexus_7_(2012)", "Nexus_7_(2013)", "Nexus_10", "Nexus_9", "Comparison_of_Google_Nexus_tablets", "Google_Pixel", "Chromebook_Pixel", "Pixelbook", "Pixelbook_Go", "Pixel_C", "Pixel_Slate", "Chromebit", "Chromebox", "Google_Clips", "Chromecast", "Nexus_Player", "Nexus_Q", "Dropcam", "Fitbit", "List_of_Fitbit_products", "Google_Nest", "Google_Nest#Nest_Cam_Indoor", "Google_Nest#Nest_Doorbell", "Google_Nest#Nest_Protect", "Google_Nest#Nest_Secure", "Google_Nest_(smart_speakers)", "Nest_Thermostat", "Google_Nest_Wifi", "Google_OnHub", "Pixel_Buds", "Google_Pixel#Pixel_Stand", "Pixel_Visual_Core", "Google_Search_Appliance", "Sycamore_processor", "Tensor_Processing_Unit", "Google_Cardboard", "Google_Contact_Lens", "Google_Daydream", "Google_Glass", "Don%27t_be_evil", "Gayglers", "Google_(verb)", "Googlearchy", "Google_bombing", "Google_effect", "Googlefight", "Google_hacking", "Googleshare", "Google_tax", "Googlewhack", "Googlization", "Rooting_(Android)", "YouTube_Poop", "Category:Documentary_films_about_Google", "Google_Behind_the_Screen", "Google_Current", "Google_Maps_Road_Trip", "Google_The_Thinking_Factory", "Google_and_the_World_Brain", "The_Creepy_Line", "Category:Books_about_Google", "Google_Hacks", "The_Google_Story", "Google_Volume_One", "Googled:_The_End_of_the_World_as_We_Know_It", "How_Google_Works", "I%27m_Feeling_Lucky_(book)", "In_the_Plex", ".google", "Google_Developer_Expert", "Google_Feud", "Google_Me_(film)", "Google_Me_(song)", "The_Billion_Dollar_Code", "The_Internship", "List_of_Google_products#Discontinued_products_and_services", "Category:Google", "Outline_of_Google", "Template_talk:NASDAQ-100", "Nasdaq-100", "Activision_Blizzard", "Adobe_Inc.", "Advanced_Micro_Devices", "Airbnb", "Align_Technology", "Amazon_(company)", "American_Electric_Power", "Amgen", "Analog_Devices", "Ansys", "Apple_Inc.", "Applied_Materials", "ASML_Holding", "Atlassian", "Autodesk", "ADP_(company)", "Baidu", "Biogen", "Booking_Holdings", "Broadcom_Inc.", "Cadence_Design_Systems", "Charter_Communications", "Cintas", "Cisco_Systems", "Cognizant", "Comcast", "Copart", "Costco", "CrowdStrike", "CSX_Corporation", "Datadog", "Dexcom", "DocuSign", "Dollar_Tree", "EBay", "Electronic_Arts", "Exelon", "Fiserv", "Fortinet", "Gilead_Sciences", "Honeywell", "Idexx_Laboratories", "Illumina,_Inc.", "Intel", "Intuit", "Intuitive_Surgical", "JD.com", "Keurig_Dr_Pepper", "KLA_Corporation", "Kraft_Heinz", "Lam_Research", "Lucid_Group", "Lululemon_Athletica", "Marriott_International", "Marvell_Technology_Group", "Match_Group", "MercadoLibre", "Meta_Platforms", "Microchip_Technology", "Micron_Technology", "Microsoft", "Moderna", "Mondelez_International", "Monster_Beverage", "NetEase", "Netflix", "Nvidia", "NXP_Semiconductors", "O%27Reilly_Auto_Parts", "Okta_(identity_management)", "Paccar", "Palo_Alto_Networks", "Paychex", "PayPal", "Peloton_(exercise_equipment_company)", "PepsiCo", "Pinduoduo", "Qualcomm", "Regeneron_Pharmaceuticals", "Ross_Stores", "Seagen", "Sirius_XM", "Skyworks_Solutions", "Splunk", "Starbucks", "Synopsys", "T-Mobile_US", "Tesla,_Inc.", "Texas_Instruments", "Verisign", "Verisk_Analytics", "Vertex_Pharmaceuticals", "Walgreens_Boots_Alliance", "Workday,_Inc.", "Xcel_Energy", "Xilinx", "Zoom_Video_Communications", "Zscaler", "ISNI_(identifier)"] };
-    // var start = 'Gamer';
-    // var end = 'Google';
-    // const result = await runDcp(start, end, linksW);
-    // console.log(result);
 
-    const runDcp = require('./dcp').runDcp;
-    const findPath = require('./web_scraper').findPath;
-    const prompt = require('prompt-sync')();
-    var start = prompt('Enter START wikipedia url: '); // https://en.wikipedia.org/wiki/Among_Us
-    var end = prompt('Enter END wikipedia url: '); // https://en.wikipedia.org/wiki/Google
-    const useDcp = prompt('Use DCP to path find? (y/n): ');
-    if (useDcp.toLowerCase() == 'y') {
-        start = start.slice(30, start.length);
-        end = end.slice(30, end.length);
-        const result = await runDcp(start, end, linksW);
-        console.log(result.map((urlEnd: string) => 'https://en.wikipedia.org/wiki/' + urlEnd));
-    } else if (useDcp.toLowerCase() == 'n') {
-        const result = await findPath(start, end, 1000, 8);
-        console.log(result);
-    }
+    const prompt = promptSync();
+    const start: string = prompt('Enter START wikipedia url: '); // https://en.wikipedia.org/wiki/Among_Us
+    const end: string = prompt('Enter END wikipedia url: '); // https://en.wikipedia.org/wiki/Google
+    const maxChecks: number = 1000;
+    const layerMax: number = 5;
+    const result = await findPath(start, end, maxChecks, layerMax);
+    console.log(result);
+    // const useDcp = prompt('Use DCP to path find? (y/n): ');
+    // if (useDcp.toLowerCase() == 'y') {
+    //     start = start.slice(30, start.length);
+    //     end = end.slice(30, end.length);
+    //     const result = await runDcp(start, end, linksW);
+    //     console.log(result.map((urlEnd: string) => 'https://en.wikipedia.org/wiki/' + urlEnd));
+    // } else if (useDcp.toLowerCase() == 'n') {
+    //     const result = await findPath(start, end, 1000, 8);
+    //     console.log(result);
+    // }
 
 }
 main();
