+++ web_scraper/src/web_scraping/page.ts
@@ -33,4 +33,4 @@ class Page {
     }
 }
 
-module.exports = Page;
+export default Page;
