+++ web_scraper/src/web_scraping/web_scraper.ts
@@ -1,15 +1,25 @@
 import cheerio, { Element } from "cheerio";
-const rp = require('request-promise');
-const Page = require('./page');
+import request from 'request-promise';
+import Page from './page';
 
+/**
+ * Short cut for console logging
+ * @param data object to print
+ */
 function print(data: any) {
     console.log(data);
 }
 
+/**
+ * Get a list of wikipedia urls that are contained in url.
+ * @param url Wikipedia page that you want to web scrape
+ * @returns Promise<string[]> where string[] is a list of urls that is contained in url obtained through web scraping
+ */
 async function webScrape(url: string): Promise<string[]> {
     const prefix: string = 'https://en.wikipedia.org';
-    const request = await rp({ url: url, timeout: 300000, json: true, headers: { 'Connection': 'keep-alive' } });
-    const $ = cheerio.load(request);
+    //@ts-ignore
+    const req = await request({ url: url, timeout: 300000, json: true, headers: { 'Connection': 'keep-alive' } });
+    const $ = cheerio.load(req);
     const content = $('.mw-parser-output'); // . is class and # is id
     // print(content.html());
     const construct: string[] = [];
@@ -35,27 +45,21 @@ async function webScrape(url: string): Promise<string[]> {
     return construct;
 }
 
-async function findPath(startUrl: string, endUrl: string, maxChecks: number, layerMax: number): Promise<string[] | null> {
-    // SETTINGS ================================
-
-    // const startUrl: string = 'A'; // FAKE
-    // const start: Page = new Page(startUrl); // FAKE
-    // const startUrl: string = 'https://en.wikipedia.org/wiki/Among_Us'; // REAL
-    const start: Page = new Page(startUrl); //REAL
-
-    // const endUrl: string = 'K'; // FAKE
-    // const end: Page = new Page(endUrl); // FAKE
-    // const endUrl: string = 'https://en.wikipedia.org/wiki/Video_game_development'; // REAL
-    const end: Page = new Page(endUrl); // REAL
-
-    // const maxChecks: number = 50; // DEBUG
-    // const maxChecks: number = 1000; // REAL (for now)
-
-    // SETTINGS ================================
-
+/**
+ * Find the quickest possible path to get from startUrl to endUrl by only clicking wikipedia links
+ * @param startUrl Start wikipedia page (eg: 'https://en.wikipedia.org/wiki/My_Youth_Romantic_Comedy_Is_Wrong,_As_I_Expected')
+ * @param endUrl End wikipedia page (eg: 'https://en.wikipedia.org/wiki/Japan')
+ * @param maxChecks amount of url checks it will make per web scrape
+ * @param layerMax amount of layers it will go through before giving up
+ * @param debugMode will print web scrape console 
+ * @returns a path to get from startUrl to endUrl
+ * string[] eg: ['https://en.wikipedia.org/wiki/My_Youth_Romantic_Comedy_Is_Wrong,_As_I_Expected', 'https://en.wikipedia.org/wiki/Japan']
+ */
+async function findPath(startUrl: string, endUrl: string, maxChecks: number, layerMax: number, debugMode: boolean = false): Promise<string[] | null> {
+    const start: Page = new Page(startUrl); // Page object for first wikipedia url
+    const end: Page = new Page(endUrl);
     start.setPath([startUrl]);
     end.setPath([endUrl]);
-
 
     const layers: Page[][] = [];
     const completeRecord: string[] = [];
@@ -67,13 +71,9 @@ async function findPath(startUrl: string, endUrl: string, maxChecks: number, lay
 
     while (!isFound && index <= layerMax) {
         const layer: Page[] = layers[index];
-        // print(`Layer [${index}]: ${layer.map((page) => page.url)}`);
         if (layer.length == 0) break;
         const possibleFind: Page | undefined = layer.find((page) => page.url == end.url);
         if (possibleFind) {
-            // check current layer
-            // print(`PATH FOUND: [Clicks :${index}]`);
-            // print(possibleFind.getPath());
             foundPath = possibleFind.getPath();
             isFound = true;
         } else {
@@ -109,5 +109,4 @@ async function findPath(startUrl: string, endUrl: string, maxChecks: number, lay
     return foundPath;
 }
 
-module.exports.findPath = findPath;
-module.exports.webScrape = webScrape;
+export { findPath, webScrape };
